Tables of S3 Generic Functions and Methods

known_s3generics() tabulates "known" S3 generic functions.

s3_in_namespace() tabulates S3 methods in a loaded namespace.

Usage

known_s3generics(.arrange_by = across(Generic:Namespace))

s3_in_namespace(namespace, .arrange_by = Method)

# S3 method for info_df
print(x, ...)

Arguments

.arrange_by

<data-masking> names of columns or functions for ordering results using the syntax of dplyr arrange(). Use desc() to sort by a variable in descending order.

namespace

(unquoted) name of a namespace.

Value

A data.frame with class attributes c("info_df", "catapult", "data.frame") and columns Generic and Namespace for known_s3generics(), and Generic, Class and Method for s3_in_namespace(), as follows: -

Generic

character, the generic function name.

Namespace

character, the namespace environment of the generic or "primitive" in the case of primitive (internally implemented) generics.

Class

character, the class for the method.

Method

character, the full method name.

Details

known_s3generics() returns a data.frame of "known" S3 generic functions, based on two character vectors in the base package namespace, namely .knownS3Generics and .S3PrimitiveGenerics.

s3_in_namespace() returns a data.frame of S3 methods in a loaded namespace.

References

R Internals: 1.2.2 Namespaces.

See also

Examples

known_s3generics()
#> _______________________________
#> "Known" S3 Generic Functions: -
#> 
#>  Generic        Namespace  
#>  AIC            stats      
#>  Complex        base       
#>  Math           base       
#>  Ops            base       
#>  Summary        base       
#>  add1           stats      
#>  anova          stats      
#>  anyNA          "primitive"
#>  as.call        "primitive"
#>  as.character   "primitive"
#>  as.character   base       
#>  as.complex     "primitive"
#>  as.data.frame  base       
#>  as.double      "primitive"
#>  as.environment "primitive"
#>  as.environment base       
#>  as.integer     "primitive"
#>  as.logical     "primitive"
#>  as.matrix      base       
#>  as.numeric     "primitive"
#>  as.raw         "primitive"
#>  as.vector      base       
#>  biplot         stats      
#>  c              "primitive"
#>  cbind          base       
#>  coef           stats      
#>  confint        stats      
#>  contour        graphics   
#>  deviance       stats      
#>  df.residual    stats      
#>  dim            "primitive"
#>  dim<-          "primitive"
#>  dimnames       "primitive"
#>  dimnames<-     "primitive"
#>  drop1          stats      
#>  edit           utils      
#>  extractAIC     stats      
#>  fitted         stats      
#>  formula        stats      
#>  hist           graphics   
#>  identify       graphics   
#>  image          graphics   
#>  is.array       "primitive"
#>  is.finite      "primitive"
#>  is.infinite    "primitive"
#>  is.matrix      "primitive"
#>  is.na          "primitive"
#>  is.nan         "primitive"
#>  is.numeric     "primitive"
#>  labels         base       
#>  length         "primitive"
#>  length<-       "primitive"
#>  levels<-       "primitive"
#>  lines          graphics   
#>  log10          "primitive"
#>  log2           "primitive"
#>  logLik         stats      
#>  matrixOps      base       
#>  model.frame    stats      
#>  model.matrix   stats      
#>  names          "primitive"
#>  names<-        "primitive"
#>  pairs          graphics   
#>  plot           base       
#>  points         graphics   
#>  predict        stats      
#>  print          base       
#>  profile        stats      
#>  qqnorm         stats      
#>  rbind          base       
#>  rep            "primitive"
#>  rep            base       
#>  residuals      stats      
#>  se.contrast    stats      
#>  seq            base       
#>  seq.int        "primitive"
#>  seq.int        base       
#>  sequence       base       
#>  solve          base       
#>  str            utils      
#>  summary        base       
#>  t              base       
#>  terms          stats      
#>  text           graphics   
#>  update         stats      
#>  vcov           stats      
#>  xtfrm          "primitive"
known_s3generics(.arrange_by = across(Namespace:Generic))
#> _______________________________
#> "Known" S3 Generic Functions: -
#> 
#>  Generic        Namespace  
#>  anyNA          "primitive"
#>  as.call        "primitive"
#>  as.character   "primitive"
#>  as.complex     "primitive"
#>  as.double      "primitive"
#>  as.environment "primitive"
#>  as.integer     "primitive"
#>  as.logical     "primitive"
#>  as.numeric     "primitive"
#>  as.raw         "primitive"
#>  c              "primitive"
#>  dim            "primitive"
#>  dim<-          "primitive"
#>  dimnames       "primitive"
#>  dimnames<-     "primitive"
#>  is.array       "primitive"
#>  is.finite      "primitive"
#>  is.infinite    "primitive"
#>  is.matrix      "primitive"
#>  is.na          "primitive"
#>  is.nan         "primitive"
#>  is.numeric     "primitive"
#>  length         "primitive"
#>  length<-       "primitive"
#>  levels<-       "primitive"
#>  log10          "primitive"
#>  log2           "primitive"
#>  names          "primitive"
#>  names<-        "primitive"
#>  rep            "primitive"
#>  seq.int        "primitive"
#>  xtfrm          "primitive"
#>  Complex        base       
#>  Math           base       
#>  Ops            base       
#>  Summary        base       
#>  as.character   base       
#>  as.data.frame  base       
#>  as.environment base       
#>  as.matrix      base       
#>  as.vector      base       
#>  cbind          base       
#>  labels         base       
#>  matrixOps      base       
#>  plot           base       
#>  print          base       
#>  rbind          base       
#>  rep            base       
#>  seq            base       
#>  seq.int        base       
#>  sequence       base       
#>  solve          base       
#>  summary        base       
#>  t              base       
#>  contour        graphics   
#>  hist           graphics   
#>  identify       graphics   
#>  image          graphics   
#>  lines          graphics   
#>  pairs          graphics   
#>  points         graphics   
#>  text           graphics   
#>  AIC            stats      
#>  add1           stats      
#>  anova          stats      
#>  biplot         stats      
#>  coef           stats      
#>  confint        stats      
#>  deviance       stats      
#>  df.residual    stats      
#>  drop1          stats      
#>  extractAIC     stats      
#>  fitted         stats      
#>  formula        stats      
#>  logLik         stats      
#>  model.frame    stats      
#>  model.matrix   stats      
#>  predict        stats      
#>  profile        stats      
#>  qqnorm         stats      
#>  residuals      stats      
#>  se.contrast    stats      
#>  terms          stats      
#>  update         stats      
#>  vcov           stats      
#>  edit           utils      
#>  str            utils      

s3_in_namespace(BitsnBobs)
#> ____________________________________
#> S3 Methods in Namespace BitsnBobs: -
#> 
#>  Generic      Class           Method                   
#>  as_decdeg    decdeg          as_decdeg.decdeg         
#>  as_decdeg    default         as_decdeg.default        
#>  as_decdeg    degminsec       as_decdeg.degminsec      
#>  as_decdeg    latlon          as_decdeg.latlon         
#>  as_decdeg    list            as_decdeg.list           
#>  as_degminsec decdeg          as_degminsec.decdeg      
#>  as_degminsec default         as_degminsec.default     
#>  as_degminsec degminsec       as_degminsec.degminsec   
#>  as_degminsec latlon          as_degminsec.latlon      
#>  as_degminsec list            as_degminsec.list        
#>  latlon       default         latlon.default           
#>  latlon       matrix          latlon.matrix            
#>  print        catapult        print.catapult           
#>  print        decdeg          print.decdeg             
#>  print        degminsec       print.degminsec          
#>  print        info_df         print.info_df            
#>  print        latlon          print.latlon             
#>  print        rostido         print.rostido            
#>  print_all    MethodsFunction print_all.MethodsFunction
#>  print_all    cohens_kappa    print_all.cohens_kappa   
#>  print_all    data.frame      print_all.data.frame     
#>  print_all    default         print_all.default        
#>  print_all    htest           print_all.htest          
#>  print_all    tbl             print_all.tbl            
#>  print_all    tbl_df          print_all.tbl_df         
#>  rbind        rostido         rbind.rostido            
s3_in_namespace(BitsnBobs, .arrange_by = across(Class:Generic))
#> ____________________________________
#> S3 Methods in Namespace BitsnBobs: -
#> 
#>  Generic      Class           Method                   
#>  print_all    MethodsFunction print_all.MethodsFunction
#>  print        catapult        print.catapult           
#>  print_all    cohens_kappa    print_all.cohens_kappa   
#>  print_all    data.frame      print_all.data.frame     
#>  as_decdeg    decdeg          as_decdeg.decdeg         
#>  as_degminsec decdeg          as_degminsec.decdeg      
#>  print        decdeg          print.decdeg             
#>  as_decdeg    default         as_decdeg.default        
#>  as_degminsec default         as_degminsec.default     
#>  latlon       default         latlon.default           
#>  print_all    default         print_all.default        
#>  as_decdeg    degminsec       as_decdeg.degminsec      
#>  as_degminsec degminsec       as_degminsec.degminsec   
#>  print        degminsec       print.degminsec          
#>  print_all    htest           print_all.htest          
#>  print        info_df         print.info_df            
#>  as_decdeg    latlon          as_decdeg.latlon         
#>  as_degminsec latlon          as_degminsec.latlon      
#>  print        latlon          print.latlon             
#>  as_decdeg    list            as_decdeg.list           
#>  as_degminsec list            as_degminsec.list        
#>  latlon       matrix          latlon.matrix            
#>  print        rostido         print.rostido            
#>  rbind        rostido         rbind.rostido            
#>  print_all    tbl             print_all.tbl            
#>  print_all    tbl_df          print_all.tbl_df         

s3_in_namespace(utils) |> head(40)
#> ________________________________
#> S3 Methods in Namespace utils: -
#> 
#>  Generic        Class                  Method                       
#>  $              bibentry               $.bibentry                   
#>  $              person                 $.person                     
#>  $<-            bibentry               $<-.bibentry                 
#>  $<-            citation               $<-.citation                 
#>  $<-            person                 $<-.person                   
#>  .AtNames       default                .AtNames.default             
#>  .DollarNames   default                .DollarNames.default         
#>  .DollarNames   environment            .DollarNames.environment     
#>  .DollarNames   list                   .DollarNames.list            
#>  Ops            roman                  Ops.roman                    
#>  Summary        roman                  Summary.roman                
#>  [              bibentry               [.bibentry                   
#>  [              getAnywhere            [.getAnywhere                
#>  [              news_db                [.news_db                    
#>  [              person                 [.person                     
#>  [              roman                  [.roman                      
#>  [[             bibentry               [[.bibentry                  
#>  [[             hashtab                [[.hashtab                   
#>  [[             person                 [[.person                    
#>  [[<-           hashtab                [[<-.hashtab                 
#>  as.character   person                 as.character.person          
#>  as.character   roman                  as.character.roman           
#>  as.data.frame  bibentry               as.data.frame.bibentry       
#>  as.data.frame  person                 as.data.frame.person         
#>  as.environment hashtab                as.environment.hashtab       
#>  as.person      default                as.person.default            
#>  as.personList  default                as.personList.default        
#>  as.personList  person                 as.personList.person         
#>  c              bibentry               c.bibentry                   
#>  c              person                 c.person                     
#>  close          txtProgressBar         close.txtProgressBar         
#>  edit           data.frame             edit.data.frame              
#>  edit           default                edit.default                 
#>  edit           matrix                 edit.matrix                  
#>  edit           vignette               edit.vignette                
#>  format         MethodsFunction        format.MethodsFunction       
#>  format         aspell                 format.aspell                
#>  format         aspell_inspect_context format.aspell_inspect_context
#>  format         bibentry               format.bibentry              
#>  format         citation               format.citation              
s3_in_namespace(utils, .arrange_by = across(Class:Generic)) |> head(40)
#> ________________________________
#> S3 Methods in Namespace utils: -
#> 
#>  Generic       Class                  Method                       
#>  print         Bibtex                 print.Bibtex                 
#>  str           Date                   str.Date                     
#>  print         Latex                  print.Latex                  
#>  format        MethodsFunction        format.MethodsFunction       
#>  print         MethodsFunction        print.MethodsFunction        
#>  str           POSIXt                 str.POSIXt                   
#>  head          array                  head.array                   
#>  tail          array                  tail.array                   
#>  format        aspell                 format.aspell                
#>  print         aspell                 print.aspell                 
#>  summary       aspell                 summary.aspell               
#>  format        aspell_inspect_context format.aspell_inspect_context
#>  print         aspell_inspect_context print.aspell_inspect_context 
#>  $             bibentry               $.bibentry                   
#>  $<-           bibentry               $<-.bibentry                 
#>  [             bibentry               [.bibentry                   
#>  [[            bibentry               [[.bibentry                  
#>  as.data.frame bibentry               as.data.frame.bibentry       
#>  c             bibentry               c.bibentry                   
#>  format        bibentry               format.bibentry              
#>  print         bibentry               print.bibentry               
#>  rep           bibentry               rep.bibentry                 
#>  sort          bibentry               sort.bibentry                
#>  toBibtex      bibentry               toBibtex.bibentry            
#>  transform     bibentry               transform.bibentry           
#>  unique        bibentry               unique.bibentry              
#>  print         browseVignettes        print.browseVignettes        
#>  print         changedFiles           print.changedFiles           
#>  $<-           citation               $<-.citation                 
#>  format        citation               format.citation              
#>  print         citation               print.citation               
#>  edit          data.frame             edit.data.frame              
#>  head          data.frame             head.array                   
#>  prompt        data.frame             prompt.data.frame            
#>  stack         data.frame             stack.data.frame             
#>  str           data.frame             str.data.frame               
#>  tail          data.frame             tail.array                   
#>  type.convert  data.frame             type.convert.data.frame      
#>  unstack       data.frame             unstack.data.frame           
#>  .AtNames      default                .AtNames.default