Create Degrees, Minutes and Seconds Object

The function degminsec() is used to create (latitude or longitude) coordinate objects represented in degrees, minutes and seconds.

Usage

degminsec(x, ...)

# S3 method for default
degminsec(dms, ..., pointafter = c("deg", "min", "sec"))

# S3 method for degminsec
print(x, ...)

Arguments

pointafter

a character string specifying the position of the decimal point in degminsec; must be one of "deg" (default), "min", or "sec". You can specify just the initial letter.

degminsec

numeric, representing a coordinate of latitude or longitude in degrees, minutes and seconds.

Value

An object of class "degminsec", representing a coordinate of latitude or longitude in degrees, minutes and seconds as a named list with components: -

dd

degrees represented by an integer with maximum absolute value of 180.

mm

minutes represented by a positive integer with value less than 60.

ss

seconds represented by a positiveinteger with value less than 60.

Details

degminsec() is a generic S3 function. The default method works with a numeric (double) representing a coordinate of latitude or longitude in degrees, minutes and seconds. The argument pointafter indicates the position of the decimal point, which may be placed after the degrees, the minutes or the (whole) seconds, and by default assumed to be placed after the degrees.

See also

Examples

degminsec(49.32464)
#> 
#> 	 49 degrees, 32 minutes, 46.4 seconds
degminsec(49.32464, pointafter = "deg")
#> 
#> 	 49 degrees, 32 minutes, 46.4 seconds
degminsec(4932.464, pointafter = "min")
#> 
#> 	 49 degrees, 32 minutes, 46.4 seconds
degminsec(493246.4, pointafter = "sec")
#> 
#> 	 49 degrees, 32 minutes, 46.4 seconds