Manage Triodos Bank Account Transaction Data

triodos_fname() returns the name of a Triodos Bank CSV format transactions file as a character string.

most_recent_fdate() returns the name of the Triodos Bank CSV format transactions file within a specified folder incorporating the most recent date.

read_triodos_csv() reads a CSV format transactions file downloaded from the Triodos website and returns the contents as a data frame.

as_rostido() reformats a data frame containing downloaded Triodos Bank transaction data.

Usage

triodos_fname(.date, filepath = NULL)

most_recent_fname(
  filepath = NULL,
  trydate = Sys.Date(),
  earliest = as.Date("2024-02-01"),
  fun = triodos_fname
)

read_triodos_csv(filename)

as_rostido(data, dateformat = "%d/%m/%Y")

# S3 method for rostido
rbind(..., .arrange_by = NULL)

# S3 method for rostido
print(x, ..., .include = NULL, maxwidth = 65L)

Arguments

.date

Date object, the date to be incorporated into a filename string.

filepath

character string, the path to a folder in which to conduct the file search; default NULL.

trydate

Date object, the most recent date within the file name from which to start the search; default Sys.Date().

earliest

Date object, the earliest date within the file name beyond which the search is discontinued; default as.Date("2024-02-01").

fun

function, used to incorporate .date into a filename search string; default triodos_fname.

filename

character string, the name of a CSV file to be read.

data

data frame, as returned by read_triodos_csv().

dateformat

character string, passed as the format argument to as.Date(); default "%d/%m/%Y".

...

for rbind() S3 method for class "rostido", data frames of class "rostido" to be combined.

for print() S3 method for class "rostido", further arguments passed to or from other methods.

.arrange_by

a list of expressions containing names of column(s) for ordering rows of the combined "rostido" data frame e.g., exprs(Account, Code, desc(Amount)). Use desc() to sort a variable in descending order; default NULL.

x

an object used to select a method.

.include

<tidy-select> names of variables to be included or excluded when printing a "rostido" data frame containing Triodos Bank transaction data; default NULL.

maxwidth

an integer, maximum width for printing Description field; default 65L.

Value

triodos_fname()

A character string representing a filename incorporating a date, of the form "Downloadyyyymmdd.csv", with attributes "date", a "Date" object, and "filepath", corresponding to the argument of the same name (if supplied).

most_recent_fdate()

An object comprising a character string representing a filename incorporating a date as specified in fun.

read_triodos_csv()

CSV transaction file data formatted by Triodos Bank, as a dataframe.

as_rostido()

An object of class "rostido" inheriting from "data.frame" containing reformatted Triodos Bank transaction data.

Details

These functions facilitate reading, formatting and combining CSV transaction files downloaded from the Triodos website.

triodos_fname() returns a character string representing the name of a CSV format transactions file downloaded from the Triodos website by concatenating the strings "Download", a date of the form "yyyymmdd" and the extension ".csv" e.g., "Download20240401.csv"

most_recent_fdate() searches the current folder or a folder specified using filepath for a filename incorporating a date specified by the .date argument, typically the current date obtained using the default Sys.Date(). The filename incorporates the date as specified by fun, typically the default triodos_fname() as above. If no such file exists, filenames incorporating earlier dates are searched for successively until either a corresponding file is found or the search is discontinued upon reaching the date specified in the earliest argument.

read_triodos_csv() reads a transactions CSV file as downloaded from the Triodos website using read.csv() and returns the contents as a data frame. The file to be read is specified in the filename argument and is located in the folder specified in that argument's "filepath" attribute if it has one, otherwise in the current working directory.

as_rostido() reformats a data frame containing Triodos Bank transaction data obtained using read_triodos_csv(), replacing character strings in the Date field with "Date" objects, and those in the Amount and Balance fields with numeric values.

By default, if no .arrange_by argument is specified, the rbind() S3 method for class "rostido" orders the results by Date, AccountNo and Code.

By default, if no .include argument is specified, the print() S3 method for class "rostido" excludes the SortCode and ChequeNo columns from the printed output.

See also

Examples

   dnldpath <- "~/Triodos Bank/Downloads"

   triodos_fname(Sys.Date())
#> [1] "Download20240429.csv"
#> attr(,"date")
#> [1] "2024-04-29"

   ## __________________________
   ## Current account 55545372

   triodos_fname(Sys.Date(), file.path(dnldpath, 55545372))
#> [1] "Download20240429.csv"
#> attr(,"date")
#> [1] "2024-04-29"
#> attr(,"filepath")
#> [1] "~/Triodos Bank/Downloads/55545372"

if (FALSE) {
   (curacc <- file.path(dnldpath, 55545372) |>
       most_recent_fname() |>
       read_triodos_csv() |>
       as_rostido())
}

   ## __________________________
   ## Savings account 55596784

if (FALSE) {
   (savacc <- file.path(dnldpath, 55596784) |>
       most_recent_fname() |>
       read_triodos_csv() |>
       as_rostido())

   savacc |> print(.include = c(Description, Code, Amount, Balance))
}

   ## ______________
   ## All accounts

if (FALSE) {
   rbind(curacc, savacc)
   rbind(curacc, savacc, .arrange_by = exprs(desc(Amount)))
   rbind(curacc, savacc, .arrange_by = exprs(Date, Code))
   rbind(curacc, savacc, .arrange_by = exprs(Code, Amount))
   rbind(curacc, savacc, .arrange_by = exprs(Code, desc(Amount)))

   rm(curacc, savacc)
}