Test For Skewness of a Variable

Computes \(G_{1}\), the expected population skewness of the values in x using skewness(), performs a t-test of its significance and calculates a confidence interval.

Usage

skewness.test(
  x,
  alternative = c("two.sided", "less", "greater"),
  se_method = c("Cramer", "simple"),
  conf.level = 0.95
)

skew.test(
  x,
  alternative = c("two.sided", "less", "greater"),
  se_method = c("Cramer", "simple"),
  conf.level = 0.95
)

Arguments

x

a numeric vector.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.

se_method

a character string specifying the method of calculating the standard error; must be one of "Cramer" (default), or "simple". You can specify just the initial letter.

conf.level

the confidence level required; default 0.95.

Value

A list with class "htest" containing the following components: -

statistic

the value of the t-statistic.

parameter

the degrees of freedom for the t-statistic.

p.value

the p-value for the test.

conf.int

confidence interval of the skewness (95% or other specified level).

estimate

the estimate of skewness.

alternative

a character string describing the alternative hypothesis.

method

the character string "Skewness with t-test" and the standard error method used.

data.name

a character string giving the name of the data.

Details

The t-statistic is given by the estimated population skewness, \(G_{1}\), divided by its standard error, \(SE_{G_{1}}\), where: -

$$SE_{G_{1}} = \displaystyle \sqrt{\frac{6n(n - 1)}{(n-2)(n+1)(n+3)}}$$

(see e.g., Joanes and Gill, 1998; Wright and Herrington 2011), or alternatively its approximation, \(\sqrt (6 / n_x)\), and the associated probability is derived from the t-distribution with \(n_{x}-2\) degrees of freedom. The t-test is conducted according to Crawley (2012), except that the default here is a two-tailed test. The corresponding confidence interval is calculated similarly from the quantiles of the t-distribution using both the alternative and conf.level arguments.

skew.test() is an alias for skewness.test().

Note

The confidence interval is poorly described in the available literature, seems somewhat controversial and should be used with caution.

References

Crawley, Michael J. (2012) The R Book. John Wiley & Sons, Incorporated. ISBN:9780470973929. p.350-352. doi:10.1002/9781118448908

Joanes, D.N., and Gill, C.A. (1998). Comparing measures of sample skewness and kurtosis. Journal of the Royal Statistical Society. Series D (The Statistician) 47(1): 183–189. doi:10.1111/1467-9884.00122

Wright, D.B., and Herrington, J.A. (2011). Problematic standard errors and confidence intervals for skewness and kurtosis. Behavior Research Methods 43(1): 8-17. doi:10.3758/s13428-010-0044-x

See also

Examples

 ## Heights of 100 randomly selected male university students, adapted from Spiegel and Stephens
 ## (Theory and Problems of Statistics. 4th edn. McGraw-Hill. 1999. ISBN 9780071755498).
 table(heights)
#> heights
#> 61 64 67 70 73 
#>  5 18 42 27  8 
 skewness.test(heights) 
#> 
#> 	Skewness with t-test (Cramer stderr)
#> 
#> data:  heights
#> t = -0.45492, df = 98, p-value = 0.6502
#> alternative hypothesis: true skewness is not equal to 0
#> 95 percent confidence interval:
#>  -0.5888187  0.3692019
#> sample estimates:
#>   skewness 
#> -0.1098084 
#> 
 skewness.test(heights, se_method = "simple")
#> 
#> 	Skewness with t-test (simple stderr)
#> 
#> data:  heights
#> t = -0.44829, df = 98, p-value = 0.6549
#> alternative hypothesis: true skewness is not equal to 0
#> 95 percent confidence interval:
#>  -0.5959017  0.3762849
#> sample estimates:
#>   skewness 
#> -0.1098084 
#> 

 ## Litter sizes in albino rats (n = 815), data from King (1924; Litter production and
 ## the sex ratio in various strains of rats. The Anatomical Record 27(5), 337-366).
 table(litter_sizes)
#> litter_sizes
#>   1   2   3   4   5   6   7   8   9  10  11  12 
#>   7  33  58 116 125 126 121 107  56  37  25   4 
 skewness.test(litter_sizes) 
#> 
#> 	Skewness with t-test (Cramer stderr)
#> 
#> data:  litter_sizes
#> t = 2.0204, df = 813, p-value = 0.04367
#> alternative hypothesis: true skewness is not equal to 0
#> 95 percent confidence interval:
#>  0.004927511 0.341148438
#> sample estimates:
#> skewness 
#> 0.173038 
#> 
 with(skewness.test(litter_sizes), c(stderr = estimate/statistic))
#> stderr.skewness 
#>      0.08564453 

 ## Compare a range of distributions, each with the three possible alternative hypotheses
 list(
   uniform = runif(30),
   normal = rnorm(30),
   lognormal = rlnorm(30),
   poisson = rpois(30, lambda = 10),
   negbinom = rnbinom(30, mu = 4, size = 2)
 ) |>
 map(\(distrib)
     c("less", "two.sided","greater") |>
     setNames(nm = _) |>
     map(\(altern)
         with(skewness.test(distrib, altern),
             data.frame(
                 Lower = conf.int[1],
                 Upper = conf.int[2],
                 Skewness = estimate,
                 t = statistic,
                 df = parameter,
                 p = p.value,
                 sig = starsig(p.value),
                 row.names = NULL
             )
         )
     ) |>
     bind_rows(.id = "Alternative")) |>
 bind_rows(.id = "Distribution")
#>    Distribution Alternative      Lower     Upper   Skewness          t df
#> 1       uniform        less       -Inf 0.3826123 -0.3435876 -0.8048575 28
#> 2       uniform   two.sided -1.2180370 0.5308619 -0.3435876 -0.8048575 28
#> 3       uniform     greater -1.0697874       Inf -0.3435876 -0.8048575 28
#> 4        normal        less       -Inf 0.8974555  0.1712556  0.4011681 28
#> 5        normal   two.sided -0.7031938 1.0457050  0.1712556  0.4011681 28
#> 6        normal     greater -0.5549443       Inf  0.1712556  0.4011681 28
#> 7     lognormal        less       -Inf 3.1644031  2.4382033  5.7115172 28
#> 8     lognormal   two.sided  1.5637538 3.3126527  2.4382033  5.7115172 28
#> 9     lognormal     greater  1.7120034       Inf  2.4382033  5.7115172 28
#> 10      poisson        less       -Inf 0.3366272 -0.3895727 -0.9125781 28
#> 11      poisson   two.sided -1.2640221 0.4848768 -0.3895727 -0.9125781 28
#> 12      poisson     greater -1.1157725       Inf -0.3895727 -0.9125781 28
#> 13     negbinom        less       -Inf 1.7455549  1.0193550  2.3878500 28
#> 14     negbinom   two.sided  0.1449056 1.8938045  1.0193550  2.3878500 28
#> 15     negbinom     greater  0.2931552       Inf  1.0193550  2.3878500 28
#>               p sig
#> 1  2.138418e-01  NS
#> 2  4.276837e-01  NS
#> 3  7.861582e-01  NS
#> 4  6.543292e-01  NS
#> 5  6.913416e-01  NS
#> 6  3.456708e-01  NS
#> 7  9.999980e-01  NS
#> 8  3.988426e-06 ***
#> 9  1.994213e-06 ***
#> 10 1.846277e-01  NS
#> 11 3.692555e-01  NS
#> 12 8.153723e-01  NS
#> 13 9.880325e-01  NS
#> 14 2.393510e-02   *
#> 15 1.196755e-02   *