Extract, Sort Unique Values, and Paste Column from Data Frame

Extract and sort unique values of a selected column from a data frame, and optionally paste into a character string.

Usage

wizard(data, col, .collapse = NULL)

Arguments

data

a data frame, or a data frame extension (e.g. a tibble).

col

<data-masking> quoted name of character column to extract.

.collapse

an optional character string to separate the results, see paste; default NULL.

Value

A vector of the same type as col or a single character string if a value for .collapse is supplied.

Details

Useful within a piped sequence to quickly extract or review sorted, unique contents of a column and optionally collapse into a single character string using paste by providing a suitable value for .collapse.

See also

Examples

starwars <- dplyr::starwars

starwars |> wizard(homeworld)
#>  [1] "Alderaan"       "Aleen Minor"    "Bespin"         "Bestine IV"    
#>  [5] "Cato Neimoidia" "Cerea"          "Champala"       "Chandrila"     
#>  [9] "Concord Dawn"   "Corellia"       "Coruscant"      "Dathomir"      
#> [13] "Dorin"          "Endor"          "Eriadu"         "Geonosis"      
#> [17] "Glee Anselm"    "Haruun Kal"     "Iktotch"        "Iridonia"      
#> [21] "Kalee"          "Kamino"         "Kashyyyk"       "Malastare"     
#> [25] "Mirial"         "Mon Cala"       "Muunilinst"     "Naboo"         
#> [29] "Nal Hutta"      "Ojom"           "Quermia"        "Rodia"         
#> [33] "Ryloth"         "Serenno"        "Shili"          "Skako"         
#> [37] "Socorro"        "Stewjon"        "Sullust"        "Tatooine"      
#> [41] "Toydaria"       "Trandosha"      "Troiken"        "Tund"          
#> [45] "Umbara"         "Utapau"         "Vulpter"        "Zolan"         
#> [49] NA              
starwars |> wizard(homeworld, ", ")
#> [1] "Alderaan, Aleen Minor, Bespin, Bestine IV, Cato Neimoidia, Cerea, Champala, Chandrila, Concord Dawn, Corellia, Coruscant, Dathomir, Dorin, Endor, Eriadu, Geonosis, Glee Anselm, Haruun Kal, Iktotch, Iridonia, Kalee, Kamino, Kashyyyk, Malastare, Mirial, Mon Cala, Muunilinst, Naboo, Nal Hutta, Ojom, Quermia, Rodia, Ryloth, Serenno, Shili, Skako, Socorro, Stewjon, Sullust, Tatooine, Toydaria, Trandosha, Troiken, Tund, Umbara, Utapau, Vulpter, Zolan, NA"
starwars |> wizard(homeworld, "\t") |> cat()
#> Alderaan	Aleen Minor	Bespin	Bestine IV	Cato Neimoidia	Cerea	Champala	Chandrila	Concord Dawn	Corellia	Coruscant	Dathomir	Dorin	Endor	Eriadu	Geonosis	Glee Anselm	Haruun Kal	Iktotch	Iridonia	Kalee	Kamino	Kashyyyk	Malastare	Mirial	Mon Cala	Muunilinst	Naboo	Nal Hutta	Ojom	Quermia	Rodia	Ryloth	Serenno	Shili	Skako	Socorro	Stewjon	Sullust	Tatooine	Toydaria	Trandosha	Troiken	Tund	Umbara	Utapau	Vulpter	Zolan	NA
starwars |> wizard(homeworld, "\n") |> cat()
#> Alderaan
#> Aleen Minor
#> Bespin
#> Bestine IV
#> Cato Neimoidia
#> Cerea
#> Champala
#> Chandrila
#> Concord Dawn
#> Corellia
#> Coruscant
#> Dathomir
#> Dorin
#> Endor
#> Eriadu
#> Geonosis
#> Glee Anselm
#> Haruun Kal
#> Iktotch
#> Iridonia
#> Kalee
#> Kamino
#> Kashyyyk
#> Malastare
#> Mirial
#> Mon Cala
#> Muunilinst
#> Naboo
#> Nal Hutta
#> Ojom
#> Quermia
#> Rodia
#> Ryloth
#> Serenno
#> Shili
#> Skako
#> Socorro
#> Stewjon
#> Sullust
#> Tatooine
#> Toydaria
#> Trandosha
#> Troiken
#> Tund
#> Umbara
#> Utapau
#> Vulpter
#> Zolan
#> NA

rm(starwars)