'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 2.5777961 1.5446906 3.0736624 3.5845874 1.6454185 0.3306382 0.4164739
#>  [8] 0.5164085 1.0755802 0.2113758 1.4770191 1.9595746 0.4205085 1.0510908
#> [15] 0.9734629 2.7074130 1.3318349 0.5072729 1.5011779 2.5671218
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 0.6088105
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1]  0.61207173  0.35262118  0.67465522  0.72102787  0.39225186 -2.02445414
#>  [7] -1.40111103 -0.93645164  0.07026924 -3.73091084  0.32296069  0.48968515
#> [13] -1.37807322  0.04860741 -0.02726049  0.63064372  0.24915618 -0.97132555
#> [19]  0.33385643  0.61045869
bc_func(0)
#>  [1]  0.94693482  0.43482365  1.12286981  1.27664339  0.49799473 -1.10673063
#>  [7] -0.87593156 -0.66085725  0.07286024 -1.55411775  0.39002594  0.67272740
#> [13] -0.86629059  0.04982848 -0.02689554  0.99599357  0.28655762 -0.67870618
#> [19]  0.40625007  0.94278537
bc_func(1)
#>  [1]  1.57779614  0.54469063  2.07366238  2.58458744  0.64541846 -0.66936182
#>  [7] -0.58352613 -0.48359154  0.07558021 -0.78862421  0.47701910  0.95957458
#> [13] -0.57949150  0.05109080 -0.02653707  1.70741300  0.33183490 -0.49272711
#> [19]  0.50117790  1.56712185
bc_func(2)
#>  [1]  2.82251647  0.69303458  4.22370020  5.92463355  0.85370095 -0.44533920
#>  [7] -0.41327476 -0.36666115  0.07843639 -0.47766014  0.59079272  1.41996627
#> [13] -0.41158630  0.05239593 -0.02618496  3.16504259  0.38689210 -0.37133711
#> [19]  0.62676754  2.79505729
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]   0.31387377   0.24289466   0.32185417   0.32609629   0.25850783
#>  [6]  -8.88855280  -4.28106914  -2.08713113   0.06544713 -34.96165431
#> [11]   0.22988574   0.28903439  -4.14951880   0.04628301  -0.02801037
#> [16]   0.31653697   0.19223315  -2.22027212   0.23480021   0.31363002
#> 
#> $`lambda -2`
#>  [1]   0.42475583   0.29045033   0.44707539   0.46108727   0.31532110
#>  [6]  -4.07366142  -2.38266708  -1.37492247   0.06780036 -10.69075867
#> [11]   0.27080888   0.36978938  -2.32761613   0.04742607  -0.02763205
#> [16]   0.43178797   0.21811678  -1.44306221   0.27812637   0.42412878
#> 
#> $`lambda -1`
#>  [1]  0.61207173  0.35262118  0.67465522  0.72102787  0.39225186 -2.02445414
#>  [7] -1.40111103 -0.93645164  0.07026924 -3.73091084  0.32296069  0.48968515
#> [13] -1.37807322  0.04860741 -0.02726049  0.63064372  0.24915618 -0.97132555
#> [19]  0.33385643  0.61045869
#> 
#> $`lambda 0`
#>  [1]  0.94693482  0.43482365  1.12286981  1.27664339  0.49799473 -1.10673063
#>  [7] -0.87593156 -0.66085725  0.07286024 -1.55411775  0.39002594  0.67272740
#> [13] -0.86629059  0.04982848 -0.02689554  0.99599357  0.28655762 -0.67870618
#> [19]  0.40625007  0.94278537
#> 
#> $`lambda 1`
#>  [1]  1.57779614  0.54469063  2.07366238  2.58458744  0.64541846 -0.66936182
#>  [7] -0.58352613 -0.48359154  0.07558021 -0.78862421  0.47701910  0.95957458
#> [13] -0.57949150  0.05109080 -0.02653707  1.70741300  0.33183490 -0.49272711
#> [19]  0.50117790  1.56712185
#> 
#> $`lambda 2`
#>  [1]  2.82251647  0.69303458  4.22370020  5.92463355  0.85370095 -0.44533920
#>  [7] -0.41327476 -0.36666115  0.07843639 -0.47766014  0.59079272  1.41996627
#> [13] -0.41158630  0.05239593 -0.02618496  3.16504259  0.38689210 -0.37133711
#> [19]  0.62676754  2.79505729
#> 
#> $`lambda 3`
#>  [1]  5.37651342  0.89524623  9.34603972 15.01977379  1.15160302 -0.32128470
#>  [7] -0.30925414 -0.28742846  0.08143649 -0.33018526  0.74074779  2.17487807
#> [13] -0.30854753  0.05374552 -0.02583908  6.28185598  0.45412924 -0.28982187
#> [19]  0.79431902  5.30587570
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -3.8909961 -3.0679833 -1.7574245 -0.4980726  0.6088105  1.4478382  2.0754460 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)