'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 0.37127668 1.22854071 2.04982809 1.95018100 1.18857353 2.76887566
#>  [7] 2.40593901 0.91636070 1.66739795 0.65300211 1.90112444 1.94178144
#> [13] 1.91299675 0.88553852 0.50008114 0.98428458 1.37439288 0.09308295
#> [19] 2.09709914 0.20340278
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] -0.02351724
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1] -1.69340913  0.18602616  0.51215421  0.48722708  0.15865533  0.63884258
#>  [7]  0.58436187 -0.09127334  0.40026314 -0.53138862  0.47399551  0.48500898
#> [13]  0.47725996 -0.12925636 -0.99967549 -0.01596634  0.27240601 -9.74310644
#> [19]  0.52315082 -3.91635362
bc_func(0)
#>  [1] -0.99080773  0.20582705  0.71775593  0.66792219  0.17275387  1.01844134
#>  [7]  0.87794027 -0.08734521  0.51126429 -0.42617492  0.64244552  0.66360582
#> [13]  0.64867099 -0.12155933 -0.69298491 -0.01584021  0.31801209 -2.37426429
#> [19]  0.74055503 -1.59256712
bc_func(1)
#>  [1] -0.62872332  0.22854071  1.04982809  0.95018100  0.18857353  1.76887566
#>  [7]  1.40593901 -0.08363930  0.66739795 -0.34699789  0.90112444  0.94178144
#> [13]  0.91299675 -0.11446148 -0.49991886 -0.01571542  0.37439288 -0.90691705
#> [19]  1.09709914 -0.79659722
bc_func(2)
#>  [1] -0.43107681  0.25465614  1.60089760  1.40160296  0.20635352  3.33333621
#>  [7]  2.39427125 -0.08014153  0.89010795 -0.28679412  1.30713707  1.38525757
#> [13]  1.32977827 -0.10791077 -0.37495943 -0.01559193  0.44447790 -0.49566778
#> [19]  1.69891240 -0.47931365
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]   -6.17973642    0.15356628    0.29463196    0.28839117    0.13481502
#>  [6]    0.31763085    0.30939880   -0.09985762    0.26142802   -0.86377925
#> [11]    0.28482156    0.28780542    0.28571918   -0.14668340   -2.33203551
#> [16]   -0.01622262    0.20493894 -412.96949920    0.29719053  -39.27696223
#> 
#> $`lambda -2`
#>  [1]  -3.12722637   0.16872329   0.38100324   0.36853197   0.14606958
#>  [6]   0.43478266   0.41362247  -0.09543875   0.32015785  -0.67257556
#> [11]   0.36165964   0.36739213   0.36337142  -0.13760996  -1.49935104
#> [16]  -0.01609380   0.23530349 -57.20716804   0.38630743 -11.58526644
#> 
#> $`lambda -1`
#>  [1] -1.69340913  0.18602616  0.51215421  0.48722708  0.15865533  0.63884258
#>  [7]  0.58436187 -0.09127334  0.40026314 -0.53138862  0.47399551  0.48500898
#> [13]  0.47725996 -0.12925636 -0.99967549 -0.01596634  0.27240601 -9.74310644
#> [19]  0.52315082 -3.91635362
#> 
#> $`lambda 0`
#>  [1] -0.99080773  0.20582705  0.71775593  0.66792219  0.17275387  1.01844134
#>  [7]  0.87794027 -0.08734521  0.51126429 -0.42617492  0.64244552  0.66360582
#> [13]  0.64867099 -0.12155933 -0.69298491 -0.01584021  0.31801209 -2.37426429
#> [19]  0.74055503 -1.59256712
#> 
#> $`lambda 1`
#>  [1] -0.62872332  0.22854071  1.04982809  0.95018100  0.18857353  1.76887566
#>  [7]  1.40593901 -0.08363930  0.66739795 -0.34699789  0.90112444  0.94178144
#> [13]  0.91299675 -0.11446148 -0.49991886 -0.01571542  0.37439288 -0.90691705
#> [19]  1.09709914 -0.79659722
#> 
#> $`lambda 2`
#>  [1] -0.43107681  0.25465614  1.60089760  1.40160296  0.20635352  3.33333621
#>  [7]  2.39427125 -0.08014153  0.89010795 -0.28679412  1.30713707  1.38525757
#> [13]  1.32977827 -0.10791077 -0.37495943 -0.01559193  0.44447790 -0.49566778
#> [19]  1.69891240 -0.47931365
#> 
#> $`lambda 3`
#>  [1] -0.31627362  0.28475052  2.53765262  2.13897997  0.22636873  6.74268754
#>  [7]  4.30896007 -0.07683880  1.21190877 -0.24051741  1.95706164  2.10717211
#> [13]  2.00023992 -0.10185992 -0.29164638 -0.01546974  0.53205580 -0.33306450
#> [19]  2.74089154 -0.33052823
#> 
#> 
#>   lambda -3   lambda -2   lambda -1    lambda 0    lambda 1    lambda 2 
#> -4.40788019 -4.18468858 -3.31968797 -1.43862947 -0.02351724  0.82711437 
#>    lambda 3 
#>  1.57360731 
#> 
#> lambda 1 
#>        5 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)