'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 0.4888720 0.5238308 1.1436385 0.5068648 0.3540929 0.3497509 0.2783218
#>  [8] 0.6382981 1.5017130 1.3691158 0.7972205 1.7225038 0.2749008 3.4267838
#> [15] 1.8670925 0.2133038 0.5449881 0.4041834 1.3821432 1.6779231
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 1.642055
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1] -1.0455252 -0.9090133  0.1255978 -0.9729125 -1.8241179 -1.8591782
#>  [7] -2.5929632 -0.5666661  0.3340938  0.2696016 -0.2543581  0.4194498
#> [13] -2.6376755  0.7081812  0.4644079 -3.6881501 -0.8349025 -1.4741243
#> [19]  0.2764860  0.4040251
bc_func(0)
#>  [1] -0.7156546 -0.6465865  0.1342148 -0.6795109 -1.0381961 -1.0505342
#>  [7] -1.2789773 -0.4489499  0.4066064  0.3141651 -0.2266239  0.5437789
#> [13] -1.2913449  1.2316222  0.6243824 -1.5450381 -0.6069913 -0.9058865
#> [19]  0.3236353  0.5175568
bc_func(1)
#>  [1] -0.5111280 -0.4761692  0.1436385 -0.4931352 -0.6459071 -0.6502491
#>  [7] -0.7216782 -0.3617019  0.5017130  0.3691158 -0.2027795  0.7225038
#> [13] -0.7250992  2.4267838  0.8670925 -0.7866962 -0.4550119 -0.5958166
#> [19]  0.3821432  0.6779231
bc_func(2)
#>  [1] -0.3805021 -0.3628006  0.1539545 -0.3715440 -0.4373091 -0.4388372
#>  [7] -0.4612685 -0.2962878  0.6275709  0.4372390 -0.1822197  0.9835097
#> [13] -0.4622148  5.3714237  1.2430173 -0.4772508 -0.3514940 -0.4183179
#> [19]  0.4551599  0.9077130
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]  -2.5196108  -1.9856926   0.1104834  -2.2264442  -7.1747178  -7.4578316
#>  [7] -15.1276479  -0.9484307   0.2349055   0.2034486  -0.3245416   0.2681107
#> [13] -15.7120690   0.3250497   0.2821202 -34.0132279  -1.7259576  -4.7149450
#> [19]   0.2070867   0.2627727
#> 
#> $`lambda -2`
#>  [1]  -1.5920867  -1.3221659   0.1177104  -1.4461919  -3.4878210  -3.5874499
#>  [7]  -5.9546923  -0.7272214   0.2782845   0.2332591  -0.2867071   0.3314807
#> [13]  -6.1163415   0.4574209   0.3565706 -10.4893756  -1.1834336  -2.5606455
#> [19]   0.2382637   0.3224070
#> 
#> $`lambda -1`
#>  [1] -1.0455252 -0.9090133  0.1255978 -0.9729125 -1.8241179 -1.8591782
#>  [7] -2.5929632 -0.5666661  0.3340938  0.2696016 -0.2543581  0.4194498
#> [13] -2.6376755  0.7081812  0.4644079 -3.6881501 -0.8349025 -1.4741243
#> [19]  0.2764860  0.4040251
#> 
#> $`lambda 0`
#>  [1] -0.7156546 -0.6465865  0.1342148 -0.6795109 -1.0381961 -1.0505342
#>  [7] -1.2789773 -0.4489499  0.4066064  0.3141651 -0.2266239  0.5437789
#> [13] -1.2913449  1.2316222  0.6243824 -1.5450381 -0.6069913 -0.9058865
#> [19]  0.3236353  0.5175568
#> 
#> $`lambda 1`
#>  [1] -0.5111280 -0.4761692  0.1436385 -0.4931352 -0.6459071 -0.6502491
#>  [7] -0.7216782 -0.3617019  0.5017130  0.3691158 -0.2027795  0.7225038
#> [13] -0.7250992  2.4267838  0.8670925 -0.7866962 -0.4550119 -0.5958166
#> [19]  0.3821432  0.6779231
#> 
#> $`lambda 2`
#>  [1] -0.3805021 -0.3628006  0.1539545 -0.3715440 -0.4373091 -0.4388372
#>  [7] -0.4612685 -0.2962878  0.6275709  0.4372390 -0.1822197  0.9835097
#> [13] -0.4622148  5.3714237  1.2430173 -0.4772508 -0.3514940 -0.4183179
#> [19]  0.4551599  0.9077130
#> 
#> $`lambda 3`
#>  [1] -0.2943872 -0.2854205  0.1652583 -0.2899268 -0.3185344 -0.3190722
#>  [7] -0.3261468 -0.2466472  0.7955252  0.5221258 -0.1644394  1.3702341
#> [13] -0.3264085 13.0800667  1.8362497 -0.3300983 -0.2793773 -0.3113236
#> [19]  0.5467785  1.2413562
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -2.6051683 -1.7932480 -0.8029396  0.2386806  1.6420553  3.2733130  4.1104233 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)