'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 4.2407961 0.8971987 1.3498916 0.9522501 0.5703141 1.2331496 3.2785612
#>  [8] 5.2543189 1.0813274 3.4302266 8.7750146 3.1173685 2.7785661 1.0316663
#> [15] 2.8563048 1.9749405 0.4759616 0.9057177 3.2911740 1.9330113
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 1.799148
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1]  0.76419522 -0.11458026  0.25919976 -0.05014431 -0.75341962  0.18906838
#>  [7]  0.69498815  0.80968038  0.07521069  0.70847407  0.88604008  0.67921663
#> [13]  0.64010213  0.03069434  0.64989731  0.49365564 -1.10100980 -0.10409682
#> [19]  0.69615706  0.48267244
bc_func(0)
#>  [1]  1.44475101 -0.10847788  0.30002427 -0.04892760 -0.56156795  0.20957155
#>  [7]  1.18740466  1.65905038  0.07818934  1.23262632  2.17190843  1.13698923
#> [13]  1.02193499  0.03117528  1.04952876  0.68053829 -0.74241809 -0.09902765
#> [19]  1.19124434  0.65907902
bc_func(1)
#>  [1]  3.24079608 -0.10280126  0.34989156 -0.04774993 -0.42968586  0.23314960
#>  [7]  2.27856117  4.25431889  0.08132737  2.43022658  7.77501458  2.11736855
#> [13]  1.77856607  0.03166632  1.85630480  0.97494054 -0.52403839 -0.09428233
#> [19]  2.29117401  0.93301126
bc_func(2)
#>  [1]  8.49217569 -0.09751721  0.41110362 -0.04660990 -0.33737089  0.26032897
#>  [7]  4.87448168 13.30393352  0.08463445  5.38322719 38.00044044  4.35899333
#> [13]  3.36021470  0.03216770  3.57923856  1.45019507 -0.38673027 -0.08983775
#> [19]  4.91591319  1.36826626
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]  0.32896278 -0.12821032  0.19781998 -0.05270079 -1.46361806  0.15557440
#>  [7]  0.32387469  0.33103544  0.06969585  0.32507466  0.33284001  0.32233025
#> [13]  0.31779457  0.02976184  0.31902908  0.29006037 -2.75812203 -0.11530898
#> [19]  0.32398302  0.28718292
#> 
#> $`lambda -2`
#>  [1]  0.47219805 -0.12114458  0.22560750 -0.05140154 -1.03724018  0.17119496
#>  [7]  0.45348389  0.48188922  0.07238236  0.45750632  0.49350657  0.44854901
#> [13]  0.43523676  0.03022327  0.43871405  0.37180770 -1.70712109 -0.10951490
#> [19]  0.45383973  0.36618610
#> 
#> $`lambda -1`
#>  [1]  0.76419522 -0.11458026  0.25919976 -0.05014431 -0.75341962  0.18906838
#>  [7]  0.69498815  0.80968038  0.07521069  0.70847407  0.88604008  0.67921663
#> [13]  0.64010213  0.03069434  0.64989731  0.49365564 -1.10100980 -0.10409682
#> [19]  0.69615706  0.48267244
#> 
#> $`lambda 0`
#>  [1]  1.44475101 -0.10847788  0.30002427 -0.04892760 -0.56156795  0.20957155
#>  [7]  1.18740466  1.65905038  0.07818934  1.23262632  2.17190843  1.13698923
#> [13]  1.02193499  0.03117528  1.04952876  0.68053829 -0.74241809 -0.09902765
#> [19]  1.19124434  0.65907902
#> 
#> $`lambda 1`
#>  [1]  3.24079608 -0.10280126  0.34989156 -0.04774993 -0.42968586  0.23314960
#>  [7]  2.27856117  4.25431889  0.08132737  2.43022658  7.77501458  2.11736855
#> [13]  1.77856607  0.03166632  1.85630480  0.97494054 -0.52403839 -0.09428233
#> [19]  2.29117401  0.93301126
#> 
#> $`lambda 2`
#>  [1]  8.49217569 -0.09751721  0.41110362 -0.04660990 -0.33737089  0.26032897
#>  [7]  4.87448168 13.30393352  0.08463445  5.38322719 38.00044044  4.35899333
#> [13]  3.36021470  0.03216770  3.57923856  1.45019507 -0.38673027 -0.08983775
#> [19]  4.91591319  1.36826626
#> 
#> $`lambda 3`
#>  [1]  25.08932227  -0.09259530   0.48659406  -0.04550616  -0.27150021
#>  [6]   0.29173291  11.41371132  48.02017913   0.08812082  13.12053486
#> [11] 224.89461747   9.76484867   6.81724104   0.03267966   7.43436579
#> [16]   2.23434620  -0.29739197  -0.08567254  11.54980849   2.07425314
#> 
#> 
#>   lambda -3   lambda -2   lambda -1    lambda 0    lambda 1    lambda 2 
#> -3.04999290 -2.38420739 -1.26860461  0.05484873  1.79914777  3.40165044 
#>    lambda 3 
#>  4.10665008 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)