'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 0.9459890 0.5165046 5.3104277 3.1896922 1.1275222 0.8324720 0.2934017
#>  [8] 0.5602166 0.9967474 1.6662136 0.7760151 1.8121427 0.2646479 2.0647003
#> [15] 1.1717717 2.3592506 0.7625133 1.5974913 1.0539107 1.8614859
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 2.093082
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1] -0.05709471 -0.93609136  0.81169125  0.68649012  0.11309949 -0.20124161
#>  [7] -2.40829708 -0.78502378 -0.00326319  0.39983686 -0.28863471  0.44816708
#> [13] -2.77860513  0.51566822  0.14659143  0.57613659 -0.31145248  0.37401851
#> [19]  0.05115297  0.46279476
bc_func(0)
#>  [1] -0.055524301 -0.660671179  1.669672376  1.159924419  0.120022472
#>  [6] -0.183355696 -1.226212776 -0.579431735 -0.003257877  0.510553767
#> [11] -0.253583293  0.594509959 -1.329354930  0.724985106  0.158516871
#> [16]  0.858344012 -0.271135286  0.468434474  0.052507687  0.621375052
bc_func(1)
#>  [1] -0.054010965 -0.483495449  4.310427689  2.189692188  0.127522189
#>  [6] -0.167528004 -0.706598346 -0.439783372 -0.003252576  0.666213633
#> [11] -0.223984895  0.812142703 -0.735352077  1.064700348  0.171771694
#> [16]  1.359250566 -0.237486668  0.597491320  0.053910663  0.861485923
bc_func(2)
#>  [1] -0.052552373 -0.366611524 13.600321118  4.587068127  0.135653143
#>  [6] -0.153495188 -0.456957735 -0.343078665 -0.003247286  0.888133936
#> [11] -0.198900278  1.141930589 -0.464980738  1.631493764  0.186524451
#> [16]  2.283031616 -0.209286709  0.775989259  0.055363843  1.232564920
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]  -0.06041655  -2.08578040   0.33110751   0.32306187   0.10079024
#>  [6]  -0.24445643 -12.86414849  -1.56254630  -0.00327385   0.26127459
#> [11]  -0.37996009   0.27731869 -17.65012763   0.29546226   0.12615242
#> [16]   0.30794954  -0.41852570   0.25156913   0.04858096   0.28165607
#> 
#> $`lambda -2`
#>  [1] -0.058724608 -1.374224881  0.482269907  0.450855779  0.106703746
#>  [6] -0.221490702 -5.308244491 -1.093154940 -0.003268514  0.319902105
#> [11] -0.330289708  0.347740214 -6.638928382  0.382711362  0.135846910
#> [16]  0.410169903 -0.359953802  0.304073586  0.049844659  0.355705263
#> 
#> $`lambda -1`
#>  [1] -0.05709471 -0.93609136  0.81169125  0.68649012  0.11309949 -0.20124161
#>  [7] -2.40829708 -0.78502378 -0.00326319  0.39983686 -0.28863471  0.44816708
#> [13] -2.77860513  0.51566822  0.14659143  0.57613659 -0.31145248  0.37401851
#> [19]  0.05115297  0.46279476
#> 
#> $`lambda 0`
#>  [1] -0.055524301 -0.660671179  1.669672376  1.159924419  0.120022472
#>  [6] -0.183355696 -1.226212776 -0.579431735 -0.003257877  0.510553767
#> [11] -0.253583293  0.594509959 -1.329354930  0.724985106  0.158516871
#> [16]  0.858344012 -0.271135286  0.468434474  0.052507687  0.621375052
#> 
#> $`lambda 1`
#>  [1] -0.054010965 -0.483495449  4.310427689  2.189692188  0.127522189
#>  [6] -0.167528004 -0.706598346 -0.439783372 -0.003252576  0.666213633
#> [11] -0.223984895  0.812142703 -0.735352077  1.064700348  0.171771694
#> [16]  1.359250566 -0.237486668  0.597491320  0.053910663  0.861485923
#> 
#> $`lambda 2`
#>  [1] -0.052552373 -0.366611524 13.600321118  4.587068127  0.135653143
#>  [6] -0.153495188 -0.456957735 -0.343078665 -0.003247286  0.888133936
#> [11] -0.198900278  1.141930589 -0.464980738  1.631493764  0.186524451
#> [16]  2.283031616 -0.209286709  0.775989259  0.055363843  1.232564920
#> 
#> $`lambda 3`
#>  [1] -0.051146301 -0.287402830 49.585823787 10.484120975  0.144475349
#>  [6] -0.141029632 -0.324914219 -0.274726706 -0.003242008  1.208618459
#> [11] -0.177561378  1.650275024 -0.327154817  2.600597278  0.202966612
#> [16]  4.043912609 -0.185551493  1.025587844  0.056869251  1.816763473
#> 
#> 
#>   lambda -3   lambda -2   lambda -1    lambda 0    lambda 1    lambda 2 
#> -2.93811243 -2.63241531 -1.83092816 -0.08530185  2.09308250  3.54091535 
#>    lambda 3 
#>  4.13831076 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)