'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 0.7213998 1.3796114 0.4491949 1.9399817 1.0939271 0.4636231 1.7803084
#>  [8] 0.5537488 3.1202459 0.3158082 0.4274394 0.9927034 0.2200309 2.7040962
#> [15] 1.1086669 1.6607622 0.5598633 0.7351320 1.5514409 1.5908194
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 1.007187
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1] -0.386193851  0.275158226 -1.226205259  0.484531209  0.085862323
#>  [6] -1.156924611  0.438299567 -0.805873124  0.679512438 -2.166479010
#> [11] -1.339512934 -0.007350248 -3.544816479  0.630190673  0.098015827
#> [16]  0.397866847 -0.786150343 -0.360299930  0.355437920  0.371393131
bc_func(0)
#>  [1] -0.326561754  0.321801891 -0.800298458  0.662678518  0.089774087
#>  [6] -0.768683416  0.576786609 -0.591044200  1.137911811 -1.152620248
#> [11] -0.849942760 -0.007323366 -1.513987348  0.994767738  0.103158306
#> [16]  0.507276673 -0.580062657 -0.307705212  0.439184139  0.464249227
bc_func(1)
#>  [1] -0.278600176  0.379611435 -0.550805122  0.939981657  0.093927124
#>  [6] -0.536376935  0.780308402 -0.446251242  2.120245894 -0.684191812
#> [11] -0.572560602 -0.007296616 -0.779969113  1.704096209  0.108666903
#> [16]  0.660762233 -0.440136714 -0.264868006  0.551440942  0.590819396
bc_func(2)
#>  [1] -0.239791147  0.451663856 -0.399111981  1.381764414  0.098338276
#>  [6] -0.392526827  1.084749003 -0.346681156  4.367967219 -0.450132594
#> [11] -0.408647781 -0.007269996 -0.475793204  3.156068155  0.114571151
#> [16]  0.879065597 -0.343276551 -0.229790476  0.703484498  0.765353176
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]  -0.554539257   0.206390441  -3.344350224   0.287678593   0.078700986
#>  [6]  -3.011570554   0.274259790  -1.629757743   0.322360664 -10.249661545
#> [11]  -3.934968246  -0.007404406 -30.958268689   0.316475090   0.088722607
#> [16]   0.260562665  -1.566138155  -0.505706873   0.244070020   0.250536045
#> 
#> $`lambda -2`
#>  [1] -0.460766696  0.237302201 -1.977994927  0.367145963  0.082176153
#>  [6] -1.826161890  0.342246312 -1.130588869  0.448643861 -4.513294661
#> [11] -2.236660385 -0.007377261 -9.827678412  0.431620531  0.093212276
#> [16]  0.318717833 -1.095166524 -0.425207950  0.292269863  0.302426702
#> 
#> $`lambda -1`
#>  [1] -0.386193851  0.275158226 -1.226205259  0.484531209  0.085862323
#>  [6] -1.156924611  0.438299567 -0.805873124  0.679512438 -2.166479010
#> [11] -1.339512934 -0.007350248 -3.544816479  0.630190673  0.098015827
#> [16]  0.397866847 -0.786150343 -0.360299930  0.355437920  0.371393131
#> 
#> $`lambda 0`
#>  [1] -0.326561754  0.321801891 -0.800298458  0.662678518  0.089774087
#>  [6] -0.768683416  0.576786609 -0.591044200  1.137911811 -1.152620248
#> [11] -0.849942760 -0.007323366 -1.513987348  0.994767738  0.103158306
#> [16]  0.507276673 -0.580062657 -0.307705212  0.439184139  0.464249227
#> 
#> $`lambda 1`
#>  [1] -0.278600176  0.379611435 -0.550805122  0.939981657  0.093927124
#>  [6] -0.536376935  0.780308402 -0.446251242  2.120245894 -0.684191812
#> [11] -0.572560602 -0.007296616 -0.779969113  1.704096209  0.108666903
#> [16]  0.660762233 -0.440136714 -0.264868006  0.551440942  0.590819396
#> 
#> $`lambda 2`
#>  [1] -0.239791147  0.451663856 -0.399111981  1.381764414  0.098338276
#>  [6] -0.392526827  1.084749003 -0.346681156  4.367967219 -0.450132594
#> [11] -0.408647781 -0.007269996 -0.475793204  3.156068155  0.114571151
#> [16]  0.879065597 -0.343276551 -0.229790476  0.703484498  0.765353176
#> 
#> $`lambda 3`
#>  [1] -0.208190253  0.541950891 -0.303121079  2.100392298  0.103025646
#>  [6] -0.300115305  1.547561311 -0.276733254  9.792836486 -0.322834310
#> [11] -0.307301642 -0.007243505 -0.329782505  6.257573359  0.120903130
#> [16]  1.193533373 -0.274837530 -0.200906889  0.911423416  1.008632250
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -3.6940444 -2.8709855 -1.5435522 -0.1943458  1.0071866  2.0644961  2.7547578 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)