'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 0.09013861 1.34412950 0.75939402 1.44228288 4.62545926 0.46323488
#>  [7] 0.90229496 0.73065954 1.66138708 1.27212262 0.42756877 1.48878171
#> [13] 4.69599067 0.41265367 0.62944097 8.27218701 0.43236428 1.04201797
#> [19] 0.53034096 1.42521163
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 2.458435
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1] -10.09402554   0.25602407  -0.31683945   0.30665474   0.78380525
#>  [6]  -1.15873211  -0.10828504  -0.36862649   0.39809331   0.21391226
#> [11]  -1.33880504   0.32830986   0.78705239  -1.42333965  -0.58871133
#> [16]   0.87911298  -1.31286449   0.04032365  -0.88557942   0.29834982
bc_func(0)
#>  [1] -2.40640672  0.29574660 -0.27523451  0.36622719  1.53157567 -0.76952106
#>  [7] -0.10281381 -0.31380767  0.50765284  0.24068686 -0.84964013  0.39795814
#> [13]  1.54670910 -0.88514661 -0.46292320  2.11289893 -0.83848679  0.04115919
#> [19] -0.63423516  0.35432032
bc_func(1)
#>  [1] -0.90986139  0.34412950 -0.24060598  0.44228288  3.62545926 -0.53676512
#>  [7] -0.09770504 -0.26934046  0.66138708  0.27212262 -0.57243123  0.48878171
#> [13]  3.69599067 -0.58734633 -0.37055903  7.27218701 -0.56763572  0.04201797
#> [19] -0.46965904  0.42521163
bc_func(2)
#>  [1] -0.49593752  0.40334206 -0.21166036  0.54008995 10.19743668 -0.39270673
#>  [7] -0.09293190 -0.23306832  0.88010352  0.30914798 -0.40859247  0.60823550
#> [13] 10.52616418 -0.41485847 -0.30190203 33.71453900 -0.40653056  0.04290073
#> [19] -0.35936923  0.51561410
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1] -454.80794971    0.19606973   -0.42782889    0.22222992    0.32996501
#>  [6]   -3.01998667   -0.12043392   -0.52120897    0.26064474    0.17141657
#> [11]   -3.93109487    0.23231838    0.33011451   -4.41041464   -1.00330442
#> [16]    0.33274447   -3.79076783    0.03871951   -1.90133582    0.21818950
#> 
#> $`lambda -2`
#>  [1] -61.03870135   0.22324991  -0.36703306   0.25963617   0.47662992
#>  [6]  -1.83006216  -0.11414786  -0.43656923   0.31885417   0.19103303
#> [11]  -2.23500450   0.27441618   0.47732666  -2.43628752  -0.76200184
#> [16]   0.49269316  -2.17467107   0.03951065  -1.27770488   0.25384351
#> 
#> $`lambda -1`
#>  [1] -10.09402554   0.25602407  -0.31683945   0.30665474   0.78380525
#>  [6]  -1.15873211  -0.10828504  -0.36862649   0.39809331   0.21391226
#> [11]  -1.33880504   0.32830986   0.78705239  -1.42333965  -0.58871133
#> [16]   0.87911298  -1.31286449   0.04032365  -0.88557942   0.29834982
#> 
#> $`lambda 0`
#>  [1] -2.40640672  0.29574660 -0.27523451  0.36622719  1.53157567 -0.76952106
#>  [7] -0.10281381 -0.31380767  0.50765284  0.24068686 -0.84964013  0.39795814
#> [13]  1.54670910 -0.88514661 -0.46292320  2.11289893 -0.83848679  0.04115919
#> [19] -0.63423516  0.35432032
#> 
#> $`lambda 1`
#>  [1] -0.90986139  0.34412950 -0.24060598  0.44228288  3.62545926 -0.53676512
#>  [7] -0.09770504 -0.26934046  0.66138708  0.27212262 -0.57243123  0.48878171
#> [13]  3.69599067 -0.58734633 -0.37055903  7.27218701 -0.56763572  0.04201797
#> [19] -0.46965904  0.42521163
#> 
#> $`lambda 2`
#>  [1] -0.49593752  0.40334206 -0.21166036  0.54008995 10.19743668 -0.39270673
#>  [7] -0.09293190 -0.23306832  0.88010352  0.30914798 -0.40859247  0.60823550
#> [13] 10.52616418 -0.41485847 -0.30190203 33.71453900 -0.40653056  0.04290073
#> [19] -0.35936923  0.51561410
#> 
#> $`lambda 3`
#>  [1]  -0.33308921   0.47613915  -0.18735774   0.66673596  32.65370504
#>  [6]  -0.30019868  -0.08846967  -0.20330921   1.19525744   0.35289030
#> [11]  -0.30727800   0.76661383  34.18584271  -0.30991069  -0.25020602
#> [16] 188.35271011  -0.30639144   0.04380821  -0.28361183   0.63164336
#> 
#> 
#>   lambda -3   lambda -2   lambda -1    lambda 0    lambda 1    lambda 2 
#> -4.47026605 -4.43274986 -3.76773486  0.04392715  2.45843522  3.50115251 
#>    lambda 3 
#>  4.08729598 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)