'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 4.5758481 0.1737738 1.1568276 1.2953280 7.2072653 0.1973367 2.7834731
#>  [8] 4.4598666 0.9843955 6.0737440 0.7061350 0.6944177 2.7099242 1.1922582
#> [15] 2.1051913 0.9568835 1.5630208 0.7414526 1.6145185 1.3949869
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 1.468268
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1]  0.78146128 -4.75460862  0.13556696  0.22799476  0.86125112 -4.06747984
#>  [7]  0.64073660  0.77577805 -0.01585187  0.83535691 -0.41615988 -0.44005548
#> [13]  0.63098599  0.16125549  0.52498380 -0.04505935  0.36021324 -0.34870393
#> [19]  0.38062027  0.28314741
bc_func(0)
#>  [1]  1.52079207 -1.75000103  0.14568144  0.25876394  1.97508958 -1.62284362
#>  [7]  1.02369947  1.49511885 -0.01572754  1.80397522 -0.34794890 -0.36468164
#> [13]  0.99692067  0.17584913  0.74440636 -0.04407368  0.44662035 -0.29914408
#> [19]  0.47903674  0.33288505
bc_func(1)
#>  [1]  3.57584813 -0.82622624  0.15682761  0.29532799  6.20726527 -0.80266325
#>  [7]  1.78347310  3.45986658 -0.01560451  5.07374398 -0.29386504 -0.30558231
#> [13]  1.70992421  0.19225817  1.10519135 -0.04311655  0.56302079 -0.25854743
#> [19]  0.61451846  0.39498693
bc_func(2)
#>  [1]  9.96919308 -0.48490134  0.16912505  0.33893731 25.47233630 -0.48052910
#>  [7]  3.37386126  9.44520496 -0.01548276 17.94518296 -0.25068671 -0.25889204
#> [13]  3.17184460  0.21073977  1.71591530 -0.04218703  0.72151700 -0.22512404
#> [19]  0.80333492  0.47299427
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]   0.32985426 -63.18895305   0.11801906   0.17996366   0.33244297
#>  [6] -43.04319953   0.31787660   0.32957571  -0.01610448   0.33184565
#> [11]  -0.61337370  -0.66210971   0.31658362   0.13664988   0.29760572
#> [16]  -0.04712019   0.24603932  -0.48443185   0.25412886   0.21054183
#> 
#> $`lambda -2`
#>  [1]   0.47612041 -16.05776019   0.12637776   0.20200395   0.49037437
#>  [6] -12.33967599   0.43546491   0.47486226  -0.01597751   0.48644633
#> [11]  -0.50275440  -0.53687989   0.43191433   0.14825382   0.38717980
#> [16]  -0.04607452   0.29533645  -0.40950115   0.30818438   0.24306118
#> 
#> $`lambda -1`
#>  [1]  0.78146128 -4.75460862  0.13556696  0.22799476  0.86125112 -4.06747984
#>  [7]  0.64073660  0.77577805 -0.01585187  0.83535691 -0.41615988 -0.44005548
#> [13]  0.63098599  0.16125549  0.52498380 -0.04505935  0.36021324 -0.34870393
#> [19]  0.38062027  0.28314741
#> 
#> $`lambda 0`
#>  [1]  1.52079207 -1.75000103  0.14568144  0.25876394  1.97508958 -1.62284362
#>  [7]  1.02369947  1.49511885 -0.01572754  1.80397522 -0.34794890 -0.36468164
#> [13]  0.99692067  0.17584913  0.74440636 -0.04407368  0.44662035 -0.29914408
#> [19]  0.47903674  0.33288505
#> 
#> $`lambda 1`
#>  [1]  3.57584813 -0.82622624  0.15682761  0.29532799  6.20726527 -0.80266325
#>  [7]  1.78347310  3.45986658 -0.01560451  5.07374398 -0.29386504 -0.30558231
#> [13]  1.70992421  0.19225817  1.10519135 -0.04311655  0.56302079 -0.25854743
#> [19]  0.61451846  0.39498693
#> 
#> $`lambda 2`
#>  [1]  9.96919308 -0.48490134  0.16912505  0.33893731 25.47233630 -0.48052910
#>  [7]  3.37386126  9.44520496 -0.01548276 17.94518296 -0.25068671 -0.25889204
#> [13]  3.17184460  0.21073977  1.71591530 -0.04218703  0.72151700 -0.22512404
#> [19]  0.80333492  0.47299427
#> 
#> $`lambda 3`
#>  [1]  31.60362507  -0.33158417   0.18270822   0.39113265 124.45967818
#>  [6]  -0.33077178   6.85519242  29.23619151  -0.01536228  74.35421264
#> [11]  -0.21596745  -0.22171358   6.30028038   0.23159020   2.77661714
#> [16]  -0.04128423   0.93950431  -0.19746168   1.06950553   0.57154287
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -3.0877135 -2.9570551 -2.5446198 -0.4113295  1.4682684  2.2784770  2.8340827 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)