'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 0.5470683 0.7680749 0.1715001 0.2118788 2.1279451 0.3690455 0.1748652
#>  [8] 0.8466427 0.3863767 1.7820139 3.1391832 0.7256945 0.2606142 5.4146514
#> [15] 1.3523829 2.7669398 0.5579982 0.7718410 0.8862962 1.3365458
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 2.073929
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1] -0.8279254 -0.3019564 -4.8308996 -3.7196788  0.5300631 -1.7096930
#>  [7] -4.7186907 -0.1811359 -1.5881476  0.4388371  0.6814458 -0.3779903
#> [13] -2.8370901  0.8153159  0.2605644  0.6385899 -0.7921206 -0.2956036
#> [19] -0.1282910  0.2518027
bc_func(0)
#>  [1] -0.6031817 -0.2638680 -1.7631713 -1.5517407  0.7551568 -0.9968354
#>  [7] -1.7437399 -0.1664766 -0.9509424  0.5777441  1.1439626 -0.3206262
#> [13] -1.3447143  1.6891085  0.3018681  1.0177419 -0.5833996 -0.2589767
#> [19] -0.1207041  0.2900885
bc_func(1)
#>  [1] -0.4529317 -0.2319251 -0.8284999 -0.7881212  1.1279451 -0.6309545
#>  [7] -0.8251348 -0.1533573 -0.6136233  0.7820139  2.1391832 -0.2743055
#> [13] -0.7393858  4.4146514  0.3523829  1.7669398 -0.4420018 -0.2281590
#> [19] -0.1137038  0.3365458
bc_func(2)
#>  [1] -0.3503582 -0.2050305 -0.4852939 -0.4775537  1.7640751 -0.4319027
#>  [7] -0.4847111 -0.1415981 -0.4253565  1.0877867  4.4272355 -0.2366838
#> [13] -0.4660401 14.1592251  0.4144697  3.3279779 -0.3443190 -0.2021307
#> [19] -0.1072395  0.3931773
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]  -1.7025560  -0.4023112 -65.7490095 -34.7108599   0.2987396  -6.2985829
#>  [7] -62.0069195  -0.2159271  -5.4455761   0.2744292   0.3225580  -0.5388690
#> [13] -18.4981590   0.3312336   0.1985675   0.3175979  -1.5852489  -0.3915952
#> [19]  -0.1454534   0.1937199
#> 
#> $`lambda -2`
#>  [1]  -1.1706557  -0.3475452 -16.4996950 -10.6376838   0.3895796  -3.1712182
#>  [7] -15.8517114  -0.1975410  -2.8492541   0.3425481   0.4492616  -0.4494287
#> [13]  -6.8616302   0.4829459   0.2266175   0.4346914  -1.1058481  -0.3392943
#> [19]  -0.1365203   0.2201004
#> 
#> $`lambda -1`
#>  [1] -0.8279254 -0.3019564 -4.8308996 -3.7196788  0.5300631 -1.7096930
#>  [7] -4.7186907 -0.1811359 -1.5881476  0.4388371  0.6814458 -0.3779903
#> [13] -2.8370901  0.8153159  0.2605644  0.6385899 -0.7921206 -0.2956036
#> [19] -0.1282910  0.2518027
#> 
#> $`lambda 0`
#>  [1] -0.6031817 -0.2638680 -1.7631713 -1.5517407  0.7551568 -0.9968354
#>  [7] -1.7437399 -0.1664766 -0.9509424  0.5777441  1.1439626 -0.3206262
#> [13] -1.3447143  1.6891085  0.3018681  1.0177419 -0.5833996 -0.2589767
#> [19] -0.1207041  0.2900885
#> 
#> $`lambda 1`
#>  [1] -0.4529317 -0.2319251 -0.8284999 -0.7881212  1.1279451 -0.6309545
#>  [7] -0.8251348 -0.1533573 -0.6136233  0.7820139  2.1391832 -0.2743055
#> [13] -0.7393858  4.4146514  0.3523829  1.7669398 -0.4420018 -0.2281590
#> [19] -0.1137038  0.3365458
#> 
#> $`lambda 2`
#>  [1] -0.3503582 -0.2050305 -0.4852939 -0.4775537  1.7640751 -0.4319027
#>  [7] -0.4847111 -0.1415981 -0.4253565  1.0877867  4.4272355 -0.2366838
#> [13] -0.4660401 14.1592251  0.4144697  3.3279779 -0.3443190 -0.2021307
#> [19] -0.1072395  0.3931773
#> 
#> $`lambda 3`
#>  [1] -0.2787571 -0.1822942 -0.3316519 -0.3301627  2.8785516 -0.3165793
#>  [7] -0.3315510 -0.1310411 -0.3141063  1.5529720  9.9783298 -0.2059419
#> [13] -0.3274331 52.5830627  0.4911422  6.7278563 -0.2754202 -0.1800615
#> [19] -0.1012653  0.4625149
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -2.2263096 -1.8919899 -1.2554667  0.1540316  2.0739290  3.4735747  4.1226582 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)