'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 6.6767722 7.5408973 0.6799885 0.7172079 0.9336135 0.4873076 7.8681929
#>  [8] 1.0075702 2.3008147 0.3700890 1.2907916 0.5665780 0.5972844 1.6592243
#> [15] 1.7981322 0.3112701 0.3431681 0.2655950 2.3700408 0.5221798
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 1.874237
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1]  0.850227031  0.867389786 -0.470613182 -0.394295761 -0.071107068
#>  [6] -1.052091972  0.872906013  0.007513356  0.565371339 -1.702052457
#> [11]  0.225281584 -0.764982055 -0.674244305  0.397308743  0.443867379
#> [16] -2.212644108 -1.914023807 -2.765131684  0.578066335 -0.915049073
bc_func(0)
#>  [1]  1.898634671  2.020341178 -0.385679444 -0.332389457 -0.068692757
#>  [6] -0.718859747  2.062828413  0.007541723  0.833263266 -0.994011654
#> [11]  0.255255650 -0.568140523 -0.515361903  0.506350225  0.586748487
#> [16] -1.167094308 -1.069534877 -1.325782836  0.862907168 -0.649743248
bc_func(1)
#>  [1]  5.676772223  6.540897285 -0.320011535 -0.282792053 -0.066386518
#>  [6] -0.512692407  6.868192870  0.007570233  1.300814672 -0.629910960
#> [11]  0.290791570 -0.433422002 -0.402715603  0.659224335  0.798132249
#> [16] -0.688729916 -0.656831904 -0.734405040  1.370040795 -0.477820170
bc_func(2)
#>  [1] 21.789643661 27.932565934 -0.268807844 -0.242806380 -0.064182933
#>  [6] -0.381265655 30.454229517  0.007598888  2.146874078 -0.431517051
#> [11]  0.333071439 -0.339494686 -0.321625675  0.876512696  1.116639793
#> [16] -0.451555468 -0.441117829 -0.464729659  2.308546686 -0.363664112
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]   0.332213434   0.332555996  -0.726833243  -0.570198516  -0.076283128
#>  [6]  -2.547175485   0.332649022   0.007457046   0.305965915  -6.242640458
#> [11]   0.178340940  -1.499401473  -1.231021386   0.260360129   0.275999120
#> [16] -10.719321267  -7.914844801 -17.458440756   0.308294662  -2.007758593
#> 
#> $`lambda -2`
#>  [1]  0.48878403  0.49120727 -0.58135156 -0.47203033 -0.07363518 -1.60554073
#>  [7]  0.49192356  0.00748513  0.40554896 -3.15054374  0.19990569 -1.05758083
#> [13] -0.90154700  0.31838162  0.34535825 -4.66054108 -3.74576737 -6.58810830
#> [19]  0.41098599 -1.33370648
#> 
#> $`lambda -1`
#>  [1]  0.850227031  0.867389786 -0.470613182 -0.394295761 -0.071107068
#>  [6] -1.052091972  0.872906013  0.007513356  0.565371339 -1.702052457
#> [11]  0.225281584 -0.764982055 -0.674244305  0.397308743  0.443867379
#> [16] -2.212644108 -1.914023807 -2.765131684  0.578066335 -0.915049073
#> 
#> $`lambda 0`
#>  [1]  1.898634671  2.020341178 -0.385679444 -0.332389457 -0.068692757
#>  [6] -0.718859747  2.062828413  0.007541723  0.833263266 -0.994011654
#> [11]  0.255255650 -0.568140523 -0.515361903  0.506350225  0.586748487
#> [16] -1.167094308 -1.069534877 -1.325782836  0.862907168 -0.649743248
#> 
#> $`lambda 1`
#>  [1]  5.676772223  6.540897285 -0.320011535 -0.282792053 -0.066386518
#>  [6] -0.512692407  6.868192870  0.007570233  1.300814672 -0.629910960
#> [11]  0.290791570 -0.433422002 -0.402715603  0.659224335  0.798132249
#> [16] -0.688729916 -0.656831904 -0.734405040  1.370040795 -0.477820170
#> 
#> $`lambda 2`
#>  [1] 21.789643661 27.932565934 -0.268807844 -0.242806380 -0.064182933
#>  [6] -0.381265655 30.454229517  0.007598888  2.146874078 -0.431517051
#> [11]  0.333071439 -0.339494686 -0.321625675  0.876512696  1.116639793
#> [16] -0.451555468 -0.441117829 -0.464729659  2.308546686 -0.363664112
#> 
#> $`lambda 3`
#>  [1]  98.881915773 142.604706181  -0.228528000  -0.210359128  -0.062076874
#>  [6]  -0.294759901 162.035898645   0.007627686   3.726644476  -0.316436807
#> [11]   0.383547727  -0.272707480  -0.262306532   1.189295575   1.604621431
#> [16]  -0.323280444  -0.319862345  -0.327088250   4.104246815  -0.285872100
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -2.2480697 -1.6274912 -0.7084654  0.7212005  1.8742367  2.1842409  2.3017197 
#> 
#> lambda -1 
#>         3 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)