'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 0.4511399 0.4136848 3.4421744 0.3893110 0.6770879 2.0310251 0.4686003
#>  [8] 0.5512862 0.3001865 0.2623173 2.6540303 0.3269601 0.2450378 1.5521903
#> [15] 2.0785469 1.2569616 2.6987271 0.4758638 1.2940854 2.7590424
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 0.8474869
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1] -1.2166074 -1.4172992  0.7094860 -1.5686408 -0.4769131  0.5076378
#>  [7] -1.1340150 -0.8139400 -2.3312620 -2.8121769  0.6232146 -2.0584776
#> [13] -3.0810033  0.3557491  0.5188947  0.2044308  0.6294549 -1.1014417
#> [19]  0.2272535  0.6375554
bc_func(0)
#>  [1] -0.7959778 -0.8826509  1.2361034 -0.9433769 -0.3899542  0.7085406
#>  [7] -0.7580052 -0.5955013 -1.2033512 -1.3382004  0.9760794 -1.1179173
#> [13] -1.4063429  0.4396670  0.7316690  0.2286974  0.9927802 -0.7426236
#> [19]  0.2578042  1.0148837
bc_func(1)
#>  [1] -0.5488601 -0.5863152  2.4421744 -0.6106890 -0.3229121  1.0310251
#>  [7] -0.5313997 -0.4487138 -0.6998135 -0.7376827  1.6540303 -0.6730399
#> [13] -0.7549622  0.5521903  1.0785469  0.2569616  1.6987271 -0.5241362
#> [19]  0.2940854  1.7590424
bc_func(2)
#>  [1] -0.3982364 -0.4144324  5.4242821 -0.4242185 -0.2707760  1.5625315
#>  [7] -0.3902069 -0.3480418 -0.4549440 -0.4655948  3.0219384 -0.4465486
#> [13] -0.4699782  0.7046473  1.6601785  0.2899763  3.1415640 -0.3867768
#> [19]  0.3373285  3.3061574
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]  -3.2969879  -4.3750303   0.3251604  -5.3158915  -0.7405166   0.2935471
#>  [7]  -2.9061157  -1.6561829 -11.9893451 -18.1337325   0.3155029  -9.2032903
#> [13] -22.3224764   0.2441992   0.2962141   0.1654867   0.3163743  -2.7600288
#> [19]   0.1795214   0.3174624
#> 
#> $`lambda -2`
#>  [1] -1.9566741 -2.4216678  0.4578008 -2.7989577 -0.5906362  0.3787897
#>  [7] -1.7770100 -1.1451891 -5.0486532 -6.7663465  0.4290164 -4.1771425
#> [13] -7.8272940  0.2924704  0.3842688  0.1835348  0.4313482 -1.7080286
#> [19]  0.2014314  0.4343170
#> 
#> $`lambda -1`
#>  [1] -1.2166074 -1.4172992  0.7094860 -1.5686408 -0.4769131  0.5076378
#>  [7] -1.1340150 -0.8139400 -2.3312620 -2.8121769  0.6232146 -2.0584776
#> [13] -3.0810033  0.3557491  0.5188947  0.2044308  0.6294549 -1.1014417
#> [19]  0.2272535  0.6375554
#> 
#> $`lambda 0`
#>  [1] -0.7959778 -0.8826509  1.2361034 -0.9433769 -0.3899542  0.7085406
#>  [7] -0.7580052 -0.5955013 -1.2033512 -1.3382004  0.9760794 -1.1179173
#> [13] -1.4063429  0.4396670  0.7316690  0.2286974  0.9927802 -0.7426236
#> [19]  0.2578042  1.0148837
#> 
#> $`lambda 1`
#>  [1] -0.5488601 -0.5863152  2.4421744 -0.6106890 -0.3229121  1.0310251
#>  [7] -0.5313997 -0.4487138 -0.6998135 -0.7376827  1.6540303 -0.6730399
#> [13] -0.7549622  0.5521903  1.0785469  0.2569616  1.6987271 -0.5241362
#> [19]  0.2940854  1.7590424
#> 
#> $`lambda 2`
#>  [1] -0.3982364 -0.4144324  5.4242821 -0.4242185 -0.2707760  1.5625315
#>  [7] -0.3902069 -0.3480418 -0.4549440 -0.4655948  3.0219384 -0.4465486
#> [13] -0.4699782  0.7046473  1.6601785  0.2899763  3.1415640 -0.3867768
#> [19]  0.3373285  3.3061574
#> 
#> $`lambda 3`
#>  [1] -0.3027269 -0.3097347 13.2616080 -0.3136650 -0.2298635  2.4593689
#>  [7] -0.2990339 -0.2774850 -0.3243165 -0.3273166  5.8982209 -0.3216823
#> [13] -0.3284290  0.9132279  2.6600215  0.3286466  6.2183917 -0.2974141
#> [19]  0.3890497  6.6675664
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -1.8514995 -1.2670186 -0.5022834  0.2162430  0.8474869  1.4888656  2.1168702 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)