'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 3.7364210 0.9573867 0.2103816 0.8354561 3.6242858 2.6194368 0.1551896
#>  [8] 2.9773556 1.6999909 0.4786796 3.7793642 2.3126763 0.5644328 0.3901928
#> [15] 3.7421792 1.6806045 0.6723803 1.9267848 1.3949068 7.9513660
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 1.657307
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1]  0.73236421 -0.04451006 -3.75326685 -0.19695100  0.72408357  0.61823854
#>  [7] -5.44373251  0.66413149  0.41176156 -1.08907987  0.73540523  0.56760054
#> [13] -0.77169008 -1.56283589  0.73277603  0.40497600 -0.48725355  0.48100068
#> [19]  0.28310624  0.87423545
bc_func(0)
#>  [1]  1.31812820 -0.04354793 -1.55883214 -0.17977749  1.28765725  0.96295932
#>  [7] -1.86310795  1.09103554  0.53062290 -0.73672372  1.32955579  0.83840543
#> [13] -0.57193394 -0.94111441  1.31966812  0.51915353 -0.39693116  0.65585270
#> [19]  0.33282762  2.07334374
bc_func(1)
#>  [1]  2.73642100 -0.04261334 -0.78961838 -0.16454391  2.62428580  1.61943676
#>  [7] -0.84481044  1.97735564  0.69999090 -0.52132036  2.77936418  1.31267632
#> [13] -0.43556719 -0.60980724  2.74217921  0.68060447 -0.32761969  0.92678478
#> [19]  0.39490683  6.95136602
bc_func(2)
#>  [1]  6.48042096 -0.04170539 -0.47786979 -0.15100656  6.06772378  2.93072448
#>  [7] -0.48795810  3.93232331  0.94498453 -0.38543290  6.64179680  2.17423588
#> [13] -0.34070780 -0.42387481  6.50195263  0.91221569 -0.27395236  1.35624980
#> [19]  0.47288253 31.11211083
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]   0.3269432  -0.0465206 -35.4643840  -0.2382873   0.3263315   0.3147871
#>  [7] -88.8515516   0.3207038   0.2654850  -2.7057589   0.3271585   0.3063849
#> [13]  -1.5203776  -5.2776779   0.3269726   0.2631099  -0.7632302   0.2867341
#> [19]   0.2105207   0.3326703
#> 
#> $`lambda -2`
#>  [1]   0.46418554  -0.04550063 -10.79677288  -0.21634585   0.46193506
#>  [6]   0.42712909 -20.26084432   0.44359617   0.32698777  -1.68212736
#> [11]   0.46499480   0.40651535  -1.06944286  -2.78406389   0.46429567
#> [16]   0.32297322  -0.60596156   0.36531985   0.24303167   0.49209164
#> 
#> $`lambda -1`
#>  [1]  0.73236421 -0.04451006 -3.75326685 -0.19695100  0.72408357  0.61823854
#>  [7] -5.44373251  0.66413149  0.41176156 -1.08907987  0.73540523  0.56760054
#> [13] -0.77169008 -1.56283589  0.73277603  0.40497600 -0.48725355  0.48100068
#> [19]  0.28310624  0.87423545
#> 
#> $`lambda 0`
#>  [1]  1.31812820 -0.04354793 -1.55883214 -0.17977749  1.28765725  0.96295932
#>  [7] -1.86310795  1.09103554  0.53062290 -0.73672372  1.32955579  0.83840543
#> [13] -0.57193394 -0.94111441  1.31966812  0.51915353 -0.39693116  0.65585270
#> [19]  0.33282762  2.07334374
#> 
#> $`lambda 1`
#>  [1]  2.73642100 -0.04261334 -0.78961838 -0.16454391  2.62428580  1.61943676
#>  [7] -0.84481044  1.97735564  0.69999090 -0.52132036  2.77936418  1.31267632
#> [13] -0.43556719 -0.60980724  2.74217921  0.68060447 -0.32761969  0.92678478
#> [19]  0.39490683  6.95136602
#> 
#> $`lambda 2`
#>  [1]  6.48042096 -0.04170539 -0.47786979 -0.15100656  6.06772378  2.93072448
#>  [7] -0.48795810  3.93232331  0.94498453 -0.38543290  6.64179680  2.17423588
#> [13] -0.34070780 -0.42387481  6.50195263  0.91221569 -0.27395236  1.35624980
#> [19]  0.47288253 31.11211083
#> 
#> $`lambda 3`
#>  [1]  17.05452765  -0.04082324  -0.33022947  -0.13895421  15.53553870
#>  [6]   5.65771055  -0.33208748   8.46440189   1.30430703  -0.29677271
#> [11]  17.66096735   3.78976133  -0.27339351  -0.31353100  17.13504105
#> [16]   1.24891734  -0.23200668   2.05106258   0.57138699 167.23964268
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -3.6728416 -3.2015562 -2.2476188 -0.4898226  1.6573065  3.5246566  4.2371051 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)