'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 2.1532362 0.3499874 1.1746999 1.2918580 1.1702769 0.2157294 5.1023846
#>  [8] 0.3581149 0.7412308 1.7757000 0.2484422 1.3401650 0.5404887 2.6562820
#> [15] 0.7331198 1.4164313 0.9308331 1.1347668 0.3667232 0.9856781
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 2.405059
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1]  0.53558276 -1.85724543  0.14871876  0.22592111  0.14550140 -3.63543710
#>  [7]  0.80401321 -1.79239982 -0.34910746  0.43684180 -3.02508064  0.25382322
#> [13] -0.85017743  0.62353395 -0.36403359  0.29400034 -0.07430642  0.11876166
#> [19] -1.72685217 -0.01452999
bc_func(0)
#>  [1]  0.76697190 -1.04985802  0.16101272  0.25608149  0.15724042 -1.53373050
#>  [7]  1.62970800 -1.02690138 -0.29944323  0.57419469 -1.39254494  0.29279274
#> [13] -0.61528154  0.97692740 -0.31044619  0.34814052 -0.07167526  0.12642716
#> [19] -1.00314789 -0.01442545
bc_func(1)
#>  [1]  1.15323617 -0.65001256  0.17469992  0.29185799  0.17027694 -0.78427061
#>  [7]  4.10238461 -0.64188509 -0.25876920  0.77569996 -0.75155777  0.34016500
#> [13] -0.45951130  1.65628199 -0.26688022  0.41643127 -0.06916688  0.13476679
#> [19] -0.63327678 -0.01432190
bc_func(2)
#>  [1]  1.81821300 -0.43875440  0.18995995  0.33444854  0.18477405 -0.47673042
#>  [7] 12.51716437 -0.43587686 -0.22528845  1.07655518 -0.46913823  0.39802111
#> [13] -0.35393598  3.02791701 -0.23126770  0.50313877 -0.06677485  0.14384784
#> [19] -0.43275704 -0.01421934
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]   0.29994431  -7.44204241   0.12769791   0.17872446   0.12535753
#>  [6] -32.86764043   0.33082400  -6.92457619  -0.48516608   0.27379866
#> [11] -21.40381159   0.19484795  -1.77781564   0.31554824  -0.51263468
#> [16]   0.21603490  -0.07996462   0.10521568  -6.42537228  -0.01474214
#> 
#> $`lambda -2`
#>  [1]   0.39215831  -3.58192572   0.13766013   0.20040094   0.13491607
#>  [6] -10.24363854   0.48079459  -3.39874838  -0.41004546   0.34142642
#> [11]  -7.60063706   0.22161011  -1.21157826   0.42913666  -0.43029382
#> [16]   0.25078224  -0.07706714   0.11170950  -3.21786139  -0.01463555
#> 
#> $`lambda -1`
#>  [1]  0.53558276 -1.85724543  0.14871876  0.22592111  0.14550140 -3.63543710
#>  [7]  0.80401321 -1.79239982 -0.34910746  0.43684180 -3.02508064  0.25382322
#> [13] -0.85017743  0.62353395 -0.36403359  0.29400034 -0.07430642  0.11876166
#> [19] -1.72685217 -0.01452999
#> 
#> $`lambda 0`
#>  [1]  0.76697190 -1.04985802  0.16101272  0.25608149  0.15724042 -1.53373050
#>  [7]  1.62970800 -1.02690138 -0.29944323  0.57419469 -1.39254494  0.29279274
#> [13] -0.61528154  0.97692740 -0.31044619  0.34814052 -0.07167526  0.12642716
#> [19] -1.00314789 -0.01442545
#> 
#> $`lambda 1`
#>  [1]  1.15323617 -0.65001256  0.17469992  0.29185799  0.17027694 -0.78427061
#>  [7]  4.10238461 -0.64188509 -0.25876920  0.77569996 -0.75155777  0.34016500
#> [13] -0.45951130  1.65628199 -0.26688022  0.41643127 -0.06916688  0.13476679
#> [19] -0.63327678 -0.01432190
#> 
#> $`lambda 2`
#>  [1]  1.81821300 -0.43875440  0.18995995  0.33444854  0.18477405 -0.47673042
#>  [7] 12.51716437 -0.43587686 -0.22528845  1.07655518 -0.46913823  0.39802111
#> [13] -0.35393598  3.02791701 -0.23126770  0.50313877 -0.06677485  0.14384784
#> [19] -0.43275704 -0.01421934
#> 
#> $`lambda 3`
#>  [1]  2.99444005 -0.31904321  0.20699726  0.38532601  0.20091685 -0.32998671
#>  [7] 43.94571948 -0.31802436 -0.19758356  1.53299265 -0.32822176  0.46899765
#> [13] -0.28070270  5.91409495 -0.20199136  0.61391809 -0.06449312  0.15374476
#> [19] -0.31689363 -0.01411776
#> 
#> 
#>    lambda -3    lambda -2    lambda -1     lambda 0     lambda 1     lambda 2 
#> -2.710889654 -2.136375441 -1.353800011  0.002952923  2.405058514  3.867010349 
#>     lambda 3 
#>  4.310891931 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)