'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 1.0329570 0.1215309 1.6287828 0.4624117 3.1461768 0.1323056 1.1925704
#>  [8] 2.7223948 8.5824637 1.3563750 4.2153433 0.5030675 3.7475549 1.0668052
#> [15] 0.7441752 2.4977616 3.5616114 1.9563177 2.4100356 0.4616780
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 1.97393
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1]  0.03190553 -7.22836335  0.38604461 -1.16257520  0.68215391 -6.55825945
#>  [7]  0.16147510  0.63267634  0.88348334  0.26274077  0.76277140 -0.98780500
#> [13]  0.73315935  0.06262176 -0.34376966  0.59964153  0.71922822  0.48883559
#> [19]  0.58506836 -1.16601199
bc_func(0)
#>  [1]  0.03242561 -2.10758713  0.48783301 -0.77129974  1.14618799 -2.02264093
#>  [7]  0.17611100  1.00151192  2.14972102  0.30481571  1.43873103 -0.68703102
#> [13]  1.32110360  0.06466841 -0.29547884  0.91539496  1.27021310  0.67106400
#> [19]  0.87964150 -0.77288768
bc_func(1)
#>  [1]  0.03295704 -0.87846915  0.62878284 -0.53758834  2.14617677 -0.86769441
#>  [7]  0.19257043  1.72239476  7.58246372  0.35637501  3.21534329 -0.49693255
#> [13]  2.74755491  0.06680522 -0.25582484  1.49776156  2.56161145  0.95631773
#> [19]  1.41003555 -0.53832204
bc_func(2)
#>  [1]  0.03350013 -0.49261513  0.82646677 -0.39308773  4.44921415 -0.49124762
#>  [7]  0.21111212  3.20571662 36.32934173  0.41987659  8.38455951 -0.37346157
#> [13]  6.52208389  0.06903669 -0.22310166  2.61940641  5.84253806  1.41358953
#> [19]  2.40413568 -0.39342673
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]    0.03089840 -185.36975628    0.25619164   -3.03792787    0.32262975
#>  [6] -143.59428314    0.13680434    0.31681275    0.33280605    0.19975396
#> [11]    0.32888313   -2.28485017    0.32699997    0.05878213   -0.47548919
#> [16]    0.31194259    0.32595533    0.28881278    0.30952065   -3.05402634
#> 
#> $`lambda -2`
#>  [1]   0.03139655 -33.35298174   0.31152939  -1.83836576   0.44948693
#>  [6] -28.06364299   0.14843800   0.43253667   0.49321193   0.22822441
#> [11]   0.47186129  -1.47568436   0.46439803   0.06066102  -0.40285844
#> [16]   0.41985655   0.46058360   0.36935547   0.41391587  -1.84580397
#> 
#> $`lambda -1`
#>  [1]  0.03190553 -7.22836335  0.38604461 -1.16257520  0.68215391 -6.55825945
#>  [7]  0.16147510  0.63267634  0.88348334  0.26274077  0.76277140 -0.98780500
#> [13]  0.73315935  0.06262176 -0.34376966  0.59964153  0.71922822  0.48883559
#> [19]  0.58506836 -1.16601199
#> 
#> $`lambda 0`
#>  [1]  0.03242561 -2.10758713  0.48783301 -0.77129974  1.14618799 -2.02264093
#>  [7]  0.17611100  1.00151192  2.14972102  0.30481571  1.43873103 -0.68703102
#> [13]  1.32110360  0.06466841 -0.29547884  0.91539496  1.27021310  0.67106400
#> [19]  0.87964150 -0.77288768
#> 
#> $`lambda 1`
#>  [1]  0.03295704 -0.87846915  0.62878284 -0.53758834  2.14617677 -0.86769441
#>  [7]  0.19257043  1.72239476  7.58246372  0.35637501  3.21534329 -0.49693255
#> [13]  2.74755491  0.06680522 -0.25582484  1.49776156  2.56161145  0.95631773
#> [19]  1.41003555 -0.53832204
#> 
#> $`lambda 2`
#>  [1]  0.03350013 -0.49261513  0.82646677 -0.39308773  4.44921415 -0.49124762
#>  [7]  0.21111212  3.20571662 36.32934173  0.41987659  8.38455951 -0.37346157
#> [13]  6.52208389  0.06903669 -0.22310166  2.61940641  5.84253806  1.41358953
#> [19]  2.40413568 -0.39342673
#> 
#> $`lambda 3`
#>  [1]   0.03405514  -0.33273501   1.10701754  -0.30037501  10.04740174
#>  [6]  -0.33256134   0.23203419   6.39228235 210.39099278   0.49846508
#> [11]  24.63431220  -0.29089509  17.21042995   0.07136754  -0.19595942
#> [16]   4.86102230  14.72643745   2.16239275   4.33271350  -0.30053165
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -2.9797915 -2.8978740 -2.5141536 -0.6745464  1.9739300  3.7572886  4.3103433 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)