'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 1.5540435 3.7732118 0.5205280 2.6897626 1.4829768 4.0932829 0.9627062
#>  [8] 1.4816525 1.9450390 2.2678516 0.5898784 0.1764831 0.4568511 1.1172133
#> [15] 0.9503304 2.0604904 1.2453246 0.2082908 0.4199268 4.6674665
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 1.090394
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1]  0.35651736  0.73497380 -0.92112637  0.62821998  0.32568064  0.75569732
#>  [7] -0.03873855  0.32507793  0.48587149  0.55905405 -0.69526477 -4.66626386
#> [13] -1.18889678  0.10491577 -0.05226563  0.51467864  0.19699652 -3.80097997
#> [19] -1.38136723  0.78575101
bc_func(0)
#>  [1]  0.44086022  1.32792657 -0.65291167  0.98945295  0.39405145  1.40934731
#>  [7] -0.03800704  0.39315804  0.66528202  0.81883297 -0.52783894 -1.73452997
#> [13] -0.78339766  0.11083745 -0.05094558  0.72294402  0.21939623 -1.56882006
#> [19] -0.86767479  1.54061642
bc_func(1)
#>  [1]  0.55404347  2.77321179 -0.47947203  1.68976265  0.48297685  3.09328291
#>  [7] -0.03729384  0.48165253  0.94503899  1.26785164 -0.41012164 -0.82351687
#> [13] -0.54314885  0.11721329 -0.04966961  1.06049041  0.24532462 -0.79170919
#> [19] -0.58007317  3.66746651
bc_func(2)
#>  [1]  0.70752555  6.61856361 -0.36452532  3.11741155  0.59961017  7.87748248
#>  [7] -0.03659842  0.59764712  1.39158833  2.07157553 -0.32602176 -0.48442685
#> [13] -0.39564352  0.12408277 -0.04843608  1.62281036  0.27541670 -0.47830747
#> [19] -0.41183073 10.39262182
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]   0.24451773   0.32712828  -2.03011735   0.31620414   0.23112751
#>  [6]   0.32847303  -0.04025860   0.23085321   0.28803379   0.30475514
#> [11]  -1.29068661 -60.30805395  -3.16253116   0.09429340  -0.05504491
#> [16]   0.29522965   0.16073721 -36.55324988  -4.16817300   0.33005514
#> 
#> $`lambda -2`
#>  [1]   0.29296504   0.46488056  -1.34536327   0.43088981   0.27264670
#>  [6]   0.47015810  -0.03948888   0.27224010   0.36783594   0.40278333
#> [11]  -0.93696133 -15.55327306  -1.89563455   0.09941211  -0.05363148
#> [16]   0.38223159   0.17759271 -11.02470436  -2.33545495   0.47704868
#> 
#> $`lambda -1`
#>  [1]  0.35651736  0.73497380 -0.92112637  0.62821998  0.32568064  0.75569732
#>  [7] -0.03873855  0.32507793  0.48587149  0.55905405 -0.69526477 -4.66626386
#> [13] -1.18889678  0.10491577 -0.05226563  0.51467864  0.19699652 -3.80097997
#> [19] -1.38136723  0.78575101
#> 
#> $`lambda 0`
#>  [1]  0.44086022  1.32792657 -0.65291167  0.98945295  0.39405145  1.40934731
#>  [7] -0.03800704  0.39315804  0.66528202  0.81883297 -0.52783894 -1.73452997
#> [13] -0.78339766  0.11083745 -0.05094558  0.72294402  0.21939623 -1.56882006
#> [19] -0.86767479  1.54061642
#> 
#> $`lambda 1`
#>  [1]  0.55404347  2.77321179 -0.47947203  1.68976265  0.48297685  3.09328291
#>  [7] -0.03729384  0.48165253  0.94503899  1.26785164 -0.41012164 -0.82351687
#> [13] -0.54314885  0.11721329 -0.04966961  1.06049041  0.24532462 -0.79170919
#> [19] -0.58007317  3.66746651
#> 
#> $`lambda 2`
#>  [1]  0.70752555  6.61856361 -0.36452532  3.11741155  0.59961017  7.87748248
#>  [7] -0.03659842  0.59764712  1.39158833  2.07157553 -0.32602176 -0.48442685
#> [13] -0.39564352  0.12408277 -0.04843608  1.62281036  0.27541670 -0.47830747
#> [19] -0.41183073 10.39262182
#> 
#> $`lambda 3`
#>  [1]  0.91769813 17.57323210 -0.28632109  6.15331897  0.75379761 22.52760388
#>  [7] -0.03592030  0.75088776  2.11947538  3.55463451 -0.26491600 -0.33150107
#> [13] -0.30154975  0.13148904 -0.04724339  2.58268691  0.31043034 -0.33032110
#> [19] -0.30865024 33.56063170
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -3.1777700 -2.8535241 -2.0683866 -0.4658237  1.0903936  1.9281499  2.3746405 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)