'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1]  0.3230195  2.0407709  0.4142286  0.6090304  1.8529109  0.5719026
#>  [7]  0.8079759  0.4616952  0.4316053  1.6659393  0.3685576  0.1804571
#> [13]  0.7385282  1.5904005  0.1627686  0.3964624  0.4282867  1.1171410
#> [19]  0.1883459 13.7489624
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 4.195672
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1] -2.0957881  0.5099891 -1.4141262 -0.6419541  0.4603086 -0.7485494
#>  [7] -0.2376607 -1.1659312 -1.3169315  0.3997380 -1.7132799 -4.5414826
#> [13] -0.3540444  0.3712276 -5.1436902 -1.5223071 -1.3348847  0.1048578
#> [19] -4.3093789  0.9272672
bc_func(0)
#>  [1] -1.1300425  0.7133276 -0.8813374 -0.4958871  0.6167578 -0.5587865
#>  [7] -0.2132231 -0.7728504 -0.8402437  0.5103891 -0.9981582 -1.7122621
#> [13] -0.3030960  0.4639859 -1.8154256 -0.9251740 -0.8479625  0.1107727
#> [19] -1.6694749  2.6209634
bc_func(1)
#>  [1] -0.6769805  1.0407709 -0.5857714 -0.3909696  0.8529109 -0.4280974
#>  [7] -0.1920241 -0.5383048 -0.5683947  0.6659393 -0.6314424 -0.8195429
#> [13] -0.2614718  0.5904005 -0.8372314 -0.6035376 -0.5717133  0.1171410
#> [19] -0.8116541 12.7489624
bc_func(2)
#>  [1] -0.4478292  1.5823730 -0.4142074 -0.3145410  1.2166393 -0.3364637
#>  [7] -0.1735875 -0.3934188 -0.4068584  0.8876769 -0.4320826 -0.4837176
#> [13] -0.2272880  0.7646869 -0.4867532 -0.4214088 -0.4082853  0.1240020
#> [19] -0.4822629 94.0169834
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]  -9.55657828   0.29411438  -4.35651342  -1.14224346   0.28093528
#>  [6]  -1.44868616  -0.29861791  -3.05364727  -3.81256199   0.26123898
#> [11]  -6.32495402 -56.38933727  -0.49418467   0.25047061 -76.96438475
#> [16]  -5.01566724  -3.90968678   0.09424696 -49.55625193   0.33320508
#> 
#> $`lambda -2`
#>  [1]  -4.29195187   0.37994466  -2.41400265  -0.84800671   0.35436661
#>  [6]  -1.02871247  -0.26590203  -1.84562895  -2.18408581   0.31984279
#> [11]  -3.18094389 -14.85401477  -0.41671811   0.30232262 -18.37246475
#> [16]  -2.68101661  -2.22584340   0.09936023 -13.59475193   0.49735497
#> 
#> $`lambda -1`
#>  [1] -2.0957881  0.5099891 -1.4141262 -0.6419541  0.4603086 -0.7485494
#>  [7] -0.2376607 -1.1659312 -1.3169315  0.3997380 -1.7132799 -4.5414826
#> [13] -0.3540444  0.3712276 -5.1436902 -1.5223071 -1.3348847  0.1048578
#> [19] -4.3093789  0.9272672
#> 
#> $`lambda 0`
#>  [1] -1.1300425  0.7133276 -0.8813374 -0.4958871  0.6167578 -0.5587865
#>  [7] -0.2132231 -0.7728504 -0.8402437  0.5103891 -0.9981582 -1.7122621
#> [13] -0.3030960  0.4639859 -1.8154256 -0.9251740 -0.8479625  0.1107727
#> [19] -1.6694749  2.6209634
#> 
#> $`lambda 1`
#>  [1] -0.6769805  1.0407709 -0.5857714 -0.3909696  0.8529109 -0.4280974
#>  [7] -0.1920241 -0.5383048 -0.5683947  0.6659393 -0.6314424 -0.8195429
#> [13] -0.2614718  0.5904005 -0.8372314 -0.6035376 -0.5717133  0.1171410
#> [19] -0.8116541 12.7489624
#> 
#> $`lambda 2`
#>  [1] -0.4478292  1.5823730 -0.4142074 -0.3145410  1.2166393 -0.3364637
#>  [7] -0.1735875 -0.3934188 -0.4068584  0.8876769 -0.4320826 -0.4837176
#> [13] -0.2272880  0.7646869 -0.4867532 -0.4214088 -0.4082853  0.1240020
#> [19] -0.4822629 94.0169834
#> 
#> $`lambda 3`
#>  [1]  -0.3220985   2.4997642  -0.3096415  -0.2580332   1.7871865  -0.2709821
#>  [7]  -0.1575110  -0.3005280  -0.3065331   1.2078570  -0.3166457  -0.3313745
#> [13]  -0.1990630   1.0075725  -0.3318959  -0.3125610  -0.3071465   0.1313988
#> [19]  -0.3311062 866.0069666
#> 
#> 
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#> -2.264888 -1.959240 -1.138519  1.229400  4.195672  4.465488  4.472002 
#> 
#> lambda -1 
#>         3 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)