'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 0.52006840 1.57717049 4.59444633 1.51755227 1.55160775 1.76412773
#>  [7] 2.33656237 2.49803117 0.46205400 0.64658695 4.93890558 1.14621267
#> [13] 1.25201878 0.56307025 2.71404827 0.09465587 1.24592793 1.08013633
#> [19] 0.75821775 0.33250371
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 1.503831
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1] -0.92282399  0.36595314  0.78234592  0.34104412  0.35550721  0.43314762
#>  [7]  0.57202084  0.59968474 -1.16424921 -0.54658240  0.79752599  0.12756155
#> [13]  0.20128994 -0.77597732  0.63154672 -9.56458541  0.19738536  0.07419094
#> [19] -0.31888234 -2.00748522
bc_func(0)
#>  [1] -0.65379494  0.45563242  1.52484825  0.41709869  0.43929165  0.56765636
#>  [7]  0.84868078  0.91550289 -0.77207352 -0.43604759  1.59714376  0.13646317
#> [13]  0.22475727 -0.57435088  0.99844135 -2.35750741  0.21988058  0.07708727
#> [19] -0.27678466 -1.10110426
bc_func(1)
#>  [1] -0.47993160  0.57717049  3.59444633  0.51755227  0.55160775  0.76412773
#>  [7]  1.33656237  1.49803117 -0.53794600 -0.35341305  3.93890558  0.14621267
#> [13]  0.25201878 -0.43692975  1.71404827 -0.90534413  0.24592793  0.08013633
#> [19] -0.24178225 -0.66749629
bc_func(2)
#>  [1] -0.36476443  0.74373338 10.05446853  0.65148245  0.70374330  1.05607333
#>  [7]  2.22976185  2.62007986 -0.39325305 -0.29096266 11.69639416  0.15690174
#> [13]  0.28377552 -0.34147595  3.18302901 -0.49552013  0.27616820  0.08334725
#> [19] -0.21255292 -0.44472064
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]   -2.03638836    0.24836779    0.32989634    0.23795543    0.24409881
#>  [6]    0.27261936    0.30720290    0.31194952   -3.04576280   -0.89976565
#> [11]    0.33056648    0.11198149    0.16349090   -1.53386734    0.31665986
#> [16] -392.70609668    0.16098782    0.06882277   -0.43137690   -8.73420190
#> 
#> $`lambda -2`
#>  [1]  -1.3486261   0.2989923   0.4763134   0.2828886   0.2923145   0.3393392
#>  [7]   0.4084169   0.4198738  -1.8419873  -0.6959586   0.4795021   0.1194256
#> [13]   0.1810311  -1.0770477   0.4321211 -55.3052325   0.1779049   0.0714388
#> [19]  -0.3697253  -4.0224837
#> 
#> $`lambda -1`
#>  [1] -0.92282399  0.36595314  0.78234592  0.34104412  0.35550721  0.43314762
#>  [7]  0.57202084  0.59968474 -1.16424921 -0.54658240  0.79752599  0.12756155
#> [13]  0.20128994 -0.77597732  0.63154672 -9.56458541  0.19738536  0.07419094
#> [19] -0.31888234 -2.00748522
#> 
#> $`lambda 0`
#>  [1] -0.65379494  0.45563242  1.52484825  0.41709869  0.43929165  0.56765636
#>  [7]  0.84868078  0.91550289 -0.77207352 -0.43604759  1.59714376  0.13646317
#> [13]  0.22475727 -0.57435088  0.99844135 -2.35750741  0.21988058  0.07708727
#> [19] -0.27678466 -1.10110426
#> 
#> $`lambda 1`
#>  [1] -0.47993160  0.57717049  3.59444633  0.51755227  0.55160775  0.76412773
#>  [7]  1.33656237  1.49803117 -0.53794600 -0.35341305  3.93890558  0.14621267
#> [13]  0.25201878 -0.43692975  1.71404827 -0.90534413  0.24592793  0.08013633
#> [19] -0.24178225 -0.66749629
#> 
#> $`lambda 2`
#>  [1] -0.36476443  0.74373338 10.05446853  0.65148245  0.70374330  1.05607333
#>  [7]  2.22976185  2.62007986 -0.39325305 -0.29096266 11.69639416  0.15690174
#> [13]  0.28377552 -0.34147595  3.18302901 -0.49552013  0.27616820  0.08334725
#> [19] -0.21255292 -0.44472064
#> 
#> $`lambda 3`
#>  [1] -0.28644550  0.97438640 31.99462614  0.83162321  0.91182495  1.49674140
#>  [7]  3.91883921  4.86270449 -0.30045143 -0.24322612 39.82455944  0.16863273
#> [13]  0.32086778 -0.27382655  6.33061236 -0.33305064  0.31136643  0.08672971
#> [19] -0.18803501 -0.32107961
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -4.4678855 -4.4145984 -3.7529566 -0.7631953  1.5038312  2.4556291  2.8135163 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)