'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 0.2239222 0.1953934 5.5180783 0.3113285 0.7312914 1.8398656 0.0709345
#>  [8] 2.6208036 6.3120399 0.4246787 0.3251321 0.5852758 1.9068244 0.7335107
#> [15] 1.3774281 0.6921226 1.6855152 6.4547441 1.3449627 0.4809328
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 1.692293
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1]  -3.4658360  -4.1178806   0.8187775  -2.2120408  -0.3674440   0.4564820
#>  [7] -13.0975128   0.6184376   0.8415726  -1.3547215  -2.0756727  -0.7085962
#> [13]   0.4755678  -0.3633067   0.2740093  -0.4448306   0.4067096   0.8450752
#> [19]   0.2564850  -1.0792925
bc_func(0)
#>  [1] -1.4964564 -1.6327404  1.7080297 -1.1669065 -0.3129433  0.6096925
#>  [7] -2.6459984  0.9634810  1.8424589 -0.8564225 -1.1235236 -0.5356721
#> [13]  0.6454392 -0.3099131  0.3202180 -0.3679921  0.5220713  1.8648154
#> [19]  0.2963663 -0.7320277
bc_func(1)
#>  [1] -0.7760778 -0.8046066  4.5180783 -0.6886715 -0.2687086  0.8398656
#>  [7] -0.9290655  1.6208036  5.3120399 -0.5753213 -0.6748679 -0.4147242
#> [13]  0.9068244 -0.2664893  0.3774281 -0.3078774  0.6855152  5.4547441
#> [19]  0.3449627 -0.5190672
bc_func(2)
#>  [1] -0.4749294 -0.4809107 14.7245942 -0.4515373 -0.2326065  1.1925528
#>  [7] -0.4974841  2.9343057 19.4209241 -0.4098240 -0.4471446 -0.3287261
#> [13]  1.3179896 -0.2309810  0.4486540 -0.2604831  0.9204808 20.3318609
#> [19]  0.4044624 -0.3843518
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]  -29.3550844  -44.3503731    0.3313495  -10.7130957   -0.5189960
#>  [6]    0.2798128 -933.5792646    0.3148161    0.3320079   -4.0187520
#> [11]   -9.3650444   -1.3293023    0.2852553   -0.5112829    0.2057858
#> [16]   -0.6720451    0.2637219    0.3320938    0.1963247   -2.6632443
#> 
#> $`lambda -2`
#>  [1]  -9.4718455 -12.5963508   0.4835792  -4.6586031  -0.4349516   0.3522941
#>  [7] -98.8699329   0.4272051   0.4874504  -2.2723567  -4.2298812  -0.9596505
#> [13]   0.3624855  -0.4293026   0.2364687  -0.5437678   0.3240032   0.4879992
#> [19]   0.2235927  -1.6617288
#> 
#> $`lambda -1`
#>  [1]  -3.4658360  -4.1178806   0.8187775  -2.2120408  -0.3674440   0.4564820
#>  [7] -13.0975128   0.6184376   0.8415726  -1.3547215  -2.0756727  -0.7085962
#> [13]   0.4755678  -0.3633067   0.2740093  -0.4448306   0.4067096   0.8450752
#> [19]   0.2564850  -1.0792925
#> 
#> $`lambda 0`
#>  [1] -1.4964564 -1.6327404  1.7080297 -1.1669065 -0.3129433  0.6096925
#>  [7] -2.6459984  0.9634810  1.8424589 -0.8564225 -1.1235236 -0.5356721
#> [13]  0.6454392 -0.3099131  0.3202180 -0.3679921  0.5220713  1.8648154
#> [19]  0.2963663 -0.7320277
#> 
#> $`lambda 1`
#>  [1] -0.7760778 -0.8046066  4.5180783 -0.6886715 -0.2687086  0.8398656
#>  [7] -0.9290655  1.6208036  5.3120399 -0.5753213 -0.6748679 -0.4147242
#> [13]  0.9068244 -0.2664893  0.3774281 -0.3078774  0.6855152  5.4547441
#> [19]  0.3449627 -0.5190672
#> 
#> $`lambda 2`
#>  [1] -0.4749294 -0.4809107 14.7245942 -0.4515373 -0.2326065  1.1925528
#>  [7] -0.4974841  2.9343057 19.4209241 -0.4098240 -0.4471446 -0.3287261
#> [13]  1.3179896 -0.2309810  0.4486540 -0.2604831  0.9204808 20.3318609
#> [19]  0.4044624 -0.3843518
#> 
#> $`lambda 3`
#>  [1] -0.3295908 -0.3308467 55.6736689 -0.3232748 -0.2029716  1.7427131
#>  [7] -0.3332144  5.6670938 83.4944458 -0.3078028 -0.3218767 -0.2665050
#> [13]  1.9777245 -0.2017812  0.5378018 -0.2228166  1.2628280 89.3095545
#> [19]  0.4776455 -0.2962540
#> 
#> 
#>   lambda -3   lambda -2   lambda -1    lambda 0    lambda 1    lambda 2 
#> -4.44973334 -4.28929755 -3.13669778 -0.03330981  1.69229256  2.14128514 
#>    lambda 3 
#>  2.28367811 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)