'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1]  1.2060344  2.0042186  1.0349376  0.3226830  0.8229178  0.1744938
#>  [7]  0.5083369  0.2493510  0.2331608  0.2368719  0.1354314 16.2047779
#> [13]  2.6681125  1.7741931  0.8047796  1.9862477  1.4112275  3.7202757
#> [19]  3.1860441  2.2577325
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 3.822506
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1]  0.17083627  0.50105242  0.03375813 -2.09901669 -0.21518816 -4.73086172
#>  [7] -0.96719940 -3.01041150 -3.28888478 -3.22169190 -6.38381252  0.93828981
#> [13]  0.62520321  0.43636349 -0.24257615  0.49653813  0.29139700  0.73120272
#> [19]  0.68613115  0.55707773
bc_func(0)
#>  [1]  0.1873376  0.6952542  0.0343411 -1.1310849 -0.1948989 -1.7458659
#>  [7] -0.6766109 -1.3888939 -1.4560267 -1.4402360 -1.9992901  2.7853061
#> [13]  0.9813713  0.5733457 -0.2171868  0.6862473  0.3444599  1.3137978
#> [19]  1.1587800  0.8143610
bc_func(1)
#>  [1]  0.20603442  1.00421858  0.03493756 -0.67731700 -0.17708217 -0.82550617
#>  [7] -0.49166312 -0.75064903 -0.76683915 -0.76312814 -0.86456861 15.20477785
#> [13]  1.66811253  0.77419309 -0.19522035  0.98624775  0.41122745  2.72027570
#> [19]  2.18604407  1.25773249
bc_func(2)
#>  [1]   0.22725951   1.50844605   0.03554788  -0.44793784  -0.16140312
#>  [6]  -0.48477595  -0.37079681  -0.46891205  -0.47281801  -0.47194586
#> [11]  -0.49082917 130.79741264   3.05941223   1.07388056  -0.17616486
#> [16]   1.47259005   0.49578146   6.42022564   4.57543842   2.04867800
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]    0.14331319    0.29192922    0.03263134   -9.58755343   -0.26481560
#>  [6]  -62.40580283   -2.20427095  -21.16701791  -25.96401034  -24.74729148
#> [11] -133.85684456    0.33325500    0.31578377    0.27364684   -0.30617733
#> [16]    0.29079519    0.21473251    0.32685962    0.32302654    0.30436915
#> 
#> $`lambda -2`
#>  [1]   0.15624375   0.37552566   0.03318833  -4.30195223  -0.23834113
#>  [6] -15.92138805  -1.43493675  -7.54170018  -8.69726632  -8.41134126
#> [11] -26.76034366   0.49809593   0.42976368   0.34115694  -0.27199775
#> [16]   0.37326307   0.24894090   0.46387401   0.45074317   0.40190993
#> 
#> $`lambda -1`
#>  [1]  0.17083627  0.50105242  0.03375813 -2.09901669 -0.21518816 -4.73086172
#>  [7] -0.96719940 -3.01041150 -3.28888478 -3.22169190 -6.38381252  0.93828981
#> [13]  0.62520321  0.43636349 -0.24257615  0.49653813  0.29139700  0.73120272
#> [19]  0.68613115  0.55707773
#> 
#> $`lambda 0`
#>  [1]  0.1873376  0.6952542  0.0343411 -1.1310849 -0.1948989 -1.7458659
#>  [7] -0.6766109 -1.3888939 -1.4560267 -1.4402360 -1.9992901  2.7853061
#> [13]  0.9813713  0.5733457 -0.2171868  0.6862473  0.3444599  1.3137978
#> [19]  1.1587800  0.8143610
#> 
#> $`lambda 1`
#>  [1]  0.20603442  1.00421858  0.03493756 -0.67731700 -0.17708217 -0.82550617
#>  [7] -0.49166312 -0.75064903 -0.76683915 -0.76312814 -0.86456861 15.20477785
#> [13]  1.66811253  0.77419309 -0.19522035  0.98624775  0.41122745  2.72027570
#> [19]  2.18604407  1.25773249
#> 
#> $`lambda 2`
#>  [1]   0.22725951   1.50844605   0.03554788  -0.44793784  -0.16140312
#>  [6]  -0.48477595  -0.37079681  -0.46891205  -0.47281801  -0.47194586
#> [11]  -0.49082917 130.79741264   3.05941223   1.07388056  -0.17616486
#> [16]   1.47259005   0.49578146   6.42022564   4.57543842   2.04867800
#> 
#> $`lambda 3`
#>  [1]    0.25140000    2.35024326    0.03617241   -0.32213362   -0.14757506
#>  [6]   -0.33156233   -0.28954750   -0.32816546   -0.32910815   -0.32890318
#> [11]   -0.33250532 1418.09693643    5.99794158    1.52824534   -0.15958938
#> [16]    2.27870170    0.60351609   16.83009821   10.44704701    3.50282209
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -3.1431545 -2.3353355 -1.3419719  0.2154409  3.8225064  4.4415087  4.4707133 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)