'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 1.4496400 0.9269643 2.1263066 0.8106159 0.1815329 1.3888819 2.9889873
#>  [8] 1.1238614 0.4217223 2.0779818 1.5623998 2.2079928 1.0539870 3.1301804
#> [15] 4.2698999 0.4766816 0.1933656 0.4762711 4.9175562 0.1708566
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 1.137411
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1]  0.31017354 -0.07879016  0.52970094 -0.23362988 -4.50864385  0.27999637
#>  [7]  0.66543852  0.11021057 -1.37122870  0.51876384  0.35995896  0.54709997
#> [13]  0.05122174  0.68052959  0.76580247 -1.09783639 -4.17155185 -1.09964429
#> [19]  0.79664696 -4.85285974
bc_func(0)
#>  [1]  0.37131523 -0.07584019  0.75438650 -0.20996095 -1.70631847  0.32849903
#>  [7]  1.09493463  0.11677044 -0.86340826  0.73139715  0.44622299  0.79208386
#> [13]  0.05258016  1.14109063  1.45159038 -0.74090652 -1.64317281 -0.74176795
#> [19]  1.59281171 -1.76693039
bc_func(1)
#>  [1]  0.44963997 -0.07303567  1.12630663 -0.18938410 -0.81846712  0.38888189
#>  [7]  1.98898727  0.12386141 -0.57827771  1.07798184  0.56239982  1.20799278
#> [13]  0.05398705  2.13018036  3.26989987 -0.52331840 -0.80663444 -0.52372885
#> [19]  3.91755623 -0.82914335
bc_func(2)
#>  [1]  0.55072802 -0.07036856  1.76058993 -0.17145093 -0.48352291  0.46449645
#>  [7]  3.96702246  0.13153223 -0.41107515  1.65900425  0.72054660  1.93761606
#> [13]  0.05544435  4.39901454  8.61602244 -0.38638733 -0.48130488 -0.38658290
#> [19] 11.59117964 -0.48540400
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]   0.22391294  -0.08516109   0.29865956  -0.29246353 -55.38688747
#>  [6]   0.20891545   0.32085069   0.09851042  -4.11092275   0.29618379
#> [11]   0.24593519   0.30236728   0.04864287   0.32246481   0.32905154
#> [16]  -2.74413498 -45.77096282  -2.75209827   0.33053028 -66.49845697
#> 
#> $`lambda -2`
#>  [1]   0.26206973  -0.08189410   0.38940940  -0.26092134 -14.67257851
#>  [6]   0.24079739   0.44403431   0.10413739  -2.31136278   0.38420588
#> [11]   0.29517374   0.39744078   0.04990991   0.44896933   0.47257576
#> [16]  -1.70045876 -12.87247429  -1.70425308   0.47932377 -16.62798357
#> 
#> $`lambda -1`
#>  [1]  0.31017354 -0.07879016  0.52970094 -0.23362988 -4.50864385  0.27999637
#>  [7]  0.66543852  0.11021057 -1.37122870  0.51876384  0.35995896  0.54709997
#> [13]  0.05122174  0.68052959  0.76580247 -1.09783639 -4.17155185 -1.09964429
#> [19]  0.79664696 -4.85285974
#> 
#> $`lambda 0`
#>  [1]  0.37131523 -0.07584019  0.75438650 -0.20996095 -1.70631847  0.32849903
#>  [7]  1.09493463  0.11677044 -0.86340826  0.73139715  0.44622299  0.79208386
#> [13]  0.05258016  1.14109063  1.45159038 -0.74090652 -1.64317281 -0.74176795
#> [19]  1.59281171 -1.76693039
#> 
#> $`lambda 1`
#>  [1]  0.44963997 -0.07303567  1.12630663 -0.18938410 -0.81846712  0.38888189
#>  [7]  1.98898727  0.12386141 -0.57827771  1.07798184  0.56239982  1.20799278
#> [13]  0.05398705  2.13018036  3.26989987 -0.52331840 -0.80663444 -0.52372885
#> [19]  3.91755623 -0.82914335
#> 
#> $`lambda 2`
#>  [1]  0.55072802 -0.07036856  1.76058993 -0.17145093 -0.48352291  0.46449645
#>  [7]  3.96702246  0.13153223 -0.41107515  1.65900425  0.72054660  1.93761606
#> [13]  0.05544435  4.39901454  8.61602244 -0.38638733 -0.48130488 -0.38658290
#> [19] 11.59117964 -0.48540400
#> 
#> $`lambda 3`
#>  [1]  0.68211822 -0.06783132  2.87113823 -0.15578193 -0.33133924  0.55971444
#>  [7]  8.56791554  0.13983647 -0.30833227  2.65758108  0.93798786  3.25482578
#> [13]  0.05695410  9.88986605 25.61633534 -0.29722862 -0.33092334 -0.29732180
#> [19] 39.30603718 -0.33167078
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -2.2263222 -2.0852900 -1.6666834 -0.4559838  1.1374112  2.1678067  2.7553651 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)