'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 5.0742191 0.4416883 0.4237444 0.6900918 0.4195430 8.2890775 0.8082877
#>  [8] 3.4115076 3.0994539 5.4620643 0.8791614 0.2562547 0.4741906 1.1751448
#> [15] 1.4310336 4.6524848 1.9106491 1.1570443 0.6832139 0.2947750
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 1.539512
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1]  0.8029253 -1.2640399 -1.3599134 -0.4490825 -1.3835457  0.8793593
#>  [7] -0.2371832  0.7068745  0.6773625  0.8169190 -0.1374475 -2.9023675
#> [13] -1.1088565  0.1490411  0.3012044  0.7850611  0.4766177  0.1357289
#> [19] -0.4636704 -2.3924185
bc_func(0)
#>  [1]  1.6241726 -0.8171508 -0.8586249 -0.3709306 -0.8685892  2.1149387
#>  [7] -0.2128372  1.2271543  1.1312259  1.6978268 -0.1287867 -1.3615834
#> [13] -0.7461459  0.1613914  0.3583970  1.5374014  0.6474430  0.1458688
#> [19] -0.3809473 -1.2215431
bc_func(1)
#>  [1]  4.0742191 -0.5583117 -0.5762556 -0.3099082 -0.5804570  7.2890775
#>  [7] -0.1917123  2.4115076  2.0994539  4.4620643 -0.1208386 -0.7437453
#> [13] -0.5258094  0.1751448  0.4310336  3.6524848  0.9106491  0.1570443
#> [19] -0.3167861 -0.7052250
bc_func(2)
#>  [1] 12.3738495 -0.4024557 -0.4102204 -0.2618866 -0.4119918 33.8544029
#>  [7] -0.1733355  5.3191921  4.3033071 14.4170733 -0.1135376 -0.4671668
#> [13] -0.3875716  0.1904827  0.5239286 10.3228073  1.3252901  0.1693758
#> [19] -0.2666094 -0.4565539
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]   0.3307820  -3.5350630  -4.0476031  -0.6809472  -4.1805382   0.3327481
#>  [7]  -0.2978867   0.3249380   0.3221384   0.3312878  -0.1572049 -19.4756977
#> [13]  -2.7928888   0.1279314   0.2195891   0.3300234   0.2855435   0.1181400
#> [19]  -0.7118889 -12.6805530
#> 
#> $`lambda -2`
#>  [1]  0.4805808 -2.0629383 -2.2845957 -0.5499200 -2.3406450  0.4927229
#>  [7] -0.2653111  0.4570387  0.4479525  0.4832407 -0.1468934 -7.1142359
#> [13] -1.7236379  0.1379344  0.2558423  0.4769006  0.3630355  0.1265177
#> [19] -0.5711656 -5.2542516
#> 
#> $`lambda -1`
#>  [1]  0.8029253 -1.2640399 -1.3599134 -0.4490825 -1.3835457  0.8793593
#>  [7] -0.2371832  0.7068745  0.6773625  0.8169190 -0.1374475 -2.9023675
#> [13] -1.1088565  0.1490411  0.3012044  0.7850611  0.4766177  0.1357289
#> [19] -0.4636704 -2.3924185
#> 
#> $`lambda 0`
#>  [1]  1.6241726 -0.8171508 -0.8586249 -0.3709306 -0.8685892  2.1149387
#>  [7] -0.2128372  1.2271543  1.1312259  1.6978268 -0.1287867 -1.3615834
#> [13] -0.7461459  0.1613914  0.3583970  1.5374014  0.6474430  0.1458688
#> [19] -0.3809473 -1.2215431
#> 
#> $`lambda 1`
#>  [1]  4.0742191 -0.5583117 -0.5762556 -0.3099082 -0.5804570  7.2890775
#>  [7] -0.1917123  2.4115076  2.0994539  4.4620643 -0.1208386 -0.7437453
#> [13] -0.5258094  0.1751448  0.4310336  3.6524848  0.9106491  0.1570443
#> [19] -0.3167861 -0.7052250
#> 
#> $`lambda 2`
#>  [1] 12.3738495 -0.4024557 -0.4102204 -0.2618866 -0.4119918 33.8544029
#>  [7] -0.1733355  5.3191921  4.3033071 14.4170733 -0.1135376 -0.4671668
#> [13] -0.3875716  0.1904827  0.5239286 10.3228073  1.3252901  0.1693758
#> [19] -0.2666094 -0.4565539
#> 
#> $`lambda 3`
#>  [1]  43.2164883  -0.3046105  -0.3079709  -0.2237866  -0.3087179 189.5108724
#>  [7]  -0.1573074  12.9014787   9.5917526  53.9853424  -0.1068248  -0.3277242
#> [13]  -0.2977917   0.2076114   0.6435175  33.2352976   1.9916592   0.1829983
#> [19]  -0.2270295  -0.3247954
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -2.6766646 -1.9941512 -0.9510389  0.3509624  1.5395116  2.7104359  3.6038048 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)