'Function Factory' for Box-Cox Transformation of Data

Function factory to create functions that take \(\lambda\) as an argument for performing the Box-Cox transformation on a given dataset. Taken directly from Wickham (2019) Section 10.4.4 Exercises.

Usage

boxcox3(x)

Arguments

x

a numeric vector containing the data to be transformed.

Value

Returns a function performing Box-Cox transformations on the data x for any given value of \(\lambda\).

Details

A numeric vector containing the data to be transformed is provided as an argument to this 'function factory', which returns a function performing the Box-Cox transformation on those data for any given value of \(\lambda\). The Box-Cox transformation takes the following form: -

if \(\lambda \ne 0\) $$y(\lambda) = \displaystyle \frac{y^\lambda - 1}{\lambda}$$

if \(\lambda = 0\) $$y(\lambda) = \log(y)$$

References

Wickham, Hadley (2019) Advanced R 2nd edition. CRC Press. adv-r.hadley.nz

See also

Examples

## Create skewed data
(d <- rlnorm(20))
#>  [1] 2.0259835 9.7807096 1.1121135 2.5346691 0.3174829 0.2132956 0.6712357
#>  [8] 0.6063729 0.4004071 3.9178318 3.5502201 8.1585903 0.8068330 0.1822058
#> [15] 1.0366548 2.9610556 2.8522664 0.6391003 0.3992624 1.8327162
## Calculate skewness using BitsnBobs::skew()
d |> skew()
#> [1] 2.000001
## Box-Cox function for these data
bc_func <- boxcox3(d)

## Box-Cox transform data with various values of lambda
bc_func(-1)
#>  [1]  0.50641258  0.89775793  0.10081123  0.60547118 -2.14977617 -3.68833030
#>  [7] -0.48978967 -0.64915014 -1.49745825  0.74475678  0.71832732  0.87742980
#> [13] -0.23941388 -4.48829935  0.03535869  0.66228260  0.64940162 -0.56469974
#> [19] -1.50461824  0.45436178
bc_func(0)
#>  [1]  0.70605528  2.28041203  0.10626229  0.93006309 -1.14733139 -1.54507651
#>  [7] -0.39863495 -0.50026009 -0.91527352  1.36553839  1.26700960  2.09907139
#> [13] -0.21463859 -1.70261844  0.03599895  1.08554582  1.04811392 -0.44769395
#> [19] -0.91813632  0.60579913
bc_func(1)
#>  [1]  1.02598355  8.78070955  0.11211353  1.53466907 -0.68251712 -0.78670445
#>  [7] -0.32876431 -0.39362707 -0.59959291  2.91783179  2.55022010  7.15859027
#> [13] -0.19316701 -0.81779420  0.03665476  1.96105557  1.85226643 -0.36089975
#> [19] -0.60073756  0.83271619
bc_func(2)
#>  [1]  1.55230466 47.33113965  0.11839825  2.71227365 -0.44960231 -0.47725250
#>  [7] -0.27472132 -0.31615593 -0.41983708  7.17470296  5.80203139 32.78129757
#> [13] -0.17451027 -0.48340052  0.03732654  3.88392505  3.56771189 -0.29577543
#> [19] -0.42029475  1.17942432
## bc_func(0) same as log(d)
identical(bc_func(0), log(d))
#> [1] TRUE

seq(-3, 3, 1) |>                         ## Create a sequence from -3 to 3
  set_names(\(x) paste("lambda", x)) |>  ## Name sequence vector using rlang::set_names()
  print_lf() |>                          ## Print with line feed
  map(bc_func) |>                        ## Box-Cox transform data using each lambda value
  print_lf() |>                          ##   in sequence and print the named list
  map_dbl(skewness) |>                   ## Calculate skewness for each element of the list
  print_lf() |>                          ##   and print the numeric vector
  abs() |>                               ## Absolute skewness...
  which.min()                            ##   ...which lambda gives minimum?
#> lambda -3 lambda -2 lambda -1  lambda 0  lambda 1  lambda 2  lambda 3 
#>        -3        -2        -1         0         1         2         3 
#> 
#> $`lambda -3`
#>  [1]   0.29324934   0.33297707   0.09098984   0.31286347 -10.08307090
#>  [6] -34.01718928  -0.76884944  -1.16172910  -4.85913023   0.32779038
#> [11]   0.32588408   0.33271952  -0.30130721 -54.77180792   0.03412319
#> [16]   0.32049410   0.31896824  -0.94361045  -4.90391736   0.27918400
#> 
#> $`lambda -2`
#>  [1]   0.37818573   0.49477328   0.09572978   0.42217350  -4.46054497
#>  [6] -10.49022052  -0.60973662  -0.85984808  -2.61864886   0.46742545
#> [11]   0.46033025   0.49248827  -0.26807338 -14.56071490   0.03473357
#> [16]   0.44297348   0.43854039  -0.72414264  -2.63655627   0.35113947
#> 
#> $`lambda -1`
#>  [1]  0.50641258  0.89775793  0.10081123  0.60547118 -2.14977617 -3.68833030
#>  [7] -0.48978967 -0.64915014 -1.49745825  0.74475678  0.71832732  0.87742980
#> [13] -0.23941388 -4.48829935  0.03535869  0.66228260  0.64940162 -0.56469974
#> [19] -1.50461824  0.45436178
#> 
#> $`lambda 0`
#>  [1]  0.70605528  2.28041203  0.10626229  0.93006309 -1.14733139 -1.54507651
#>  [7] -0.39863495 -0.50026009 -0.91527352  1.36553839  1.26700960  2.09907139
#> [13] -0.21463859 -1.70261844  0.03599895  1.08554582  1.04811392 -0.44769395
#> [19] -0.91813632  0.60579913
#> 
#> $`lambda 1`
#>  [1]  1.02598355  8.78070955  0.11211353  1.53466907 -0.68251712 -0.78670445
#>  [7] -0.32876431 -0.39362707 -0.59959291  2.91783179  2.55022010  7.15859027
#> [13] -0.19316701 -0.81779420  0.03665476  1.96105557  1.85226643 -0.36089975
#> [19] -0.60073756  0.83271619
#> 
#> $`lambda 2`
#>  [1]  1.55230466 47.33113965  0.11839825  2.71227365 -0.44960231 -0.47725250
#>  [7] -0.27472132 -0.31615593 -0.41983708  7.17470296  5.80203139 32.78129757
#> [13] -0.17451027 -0.48340052  0.03732654  3.88392505  3.56771189 -0.29577543
#> [19] -0.42029475  1.17942432
#> 
#> $`lambda 3`
#>  [1]   2.43862365 311.54832291   0.12515271   5.09470046  -0.32266640
#>  [6]  -0.33009871  -0.23252327  -0.25901462  -0.31193480  19.71213013
#> [11]  14.58239901 180.68564699  -0.15825610  -0.33131699   0.03801474
#> [16]   8.32069714   7.40146539  -0.24632002  -0.31211779   1.71860544
#> 
#> 
#>  lambda -3  lambda -2  lambda -1   lambda 0   lambda 1   lambda 2   lambda 3 
#> -3.0125696 -2.5209618 -1.5340421  0.1211487  2.0000007  2.9078276  3.2410766 
#> 
#> lambda 0 
#>        4 

## Usually, lambda 0 has least absolute skewness as data were sampled from lognormal distribution

rm(d, bc_func)