Create Latitude and Longitude Object

The function latlon() is used to create latitude and longitude objects representing paired coordinates in either decimal degrees or degrees, minutes and seconds.

latlon_dd() is a convenience function, such that latlon_dd(...) is equivalent to latlon(..., decimal = TRUE).

Usage

latlon(object, ...)

# S3 method for default
latlon(object, ..., decimal = FALSE, .fmt = c("deg", "min", "sec"))

# S3 method for matrix
latlon(object, ..., decimal = FALSE, .fmt = c("deg", "min", "sec"))

latlon_dd(object)

# S3 method for latlon
print(x, ...)

Arguments

object

a numeric vector, representing a coordinate of latitude and longitude in decimal degrees or in degrees, minutes and seconds; or a two-column numeric matrix representing a number of such coordinates.

...

further arguments passed to or from other methods.

decimal

logical, if TRUE indicating whether the coordinate represented by object is in decimal degrees or otherwise in degrees, minutes and seconds; default FALSE.

.fmt

a character string indicating the position of the decimal point in object; must be one of "deg" (default), "min", or "sec". You can specify just the initial letter.

x

object to be printed.

Value

An object of class "latlon" instantiating a coordinate of latitude and longitude in decimal degrees or degrees, minutes and seconds, comprising a list of either two "decdeg" or two "degminsec" objects, with attribute "degrtype" indicating which of these two types the object is; or if length(object) > 1, a list of "latlon" objects.

Details

latlon() is a generic S3 function. The default method works with a numeric vector of length 2 representing a coordinate of paired latitude and longitude values. The method for class "matrix" works with a two column numeric matrix each row of which contains paired latitude and longitude values.

See also

Examples

## Decimal degrees
latlon(c(49.54621, 18.398562), decimal = TRUE)
#> 	49.546210, 18.398562 decimal degrees

## degrees, minutes and seconds
latlon(c(49.3246368, 18.2354822))
#> 	49°32'46.368"N, 18°23'54.822"E
latlon(c(493246.368, 182354.822), .fmt = "sec")
#> 	49°32'46.368"N, 18°23'54.822"E

## Decimal degrees—convenience function
latlon_dd(c(49.54621, 18.398562))
#> 	49.546210, 18.398562 decimal degrees

## Two-column numeric `matrix` in decimal degrees
(ll_mtx <- matrix(
       c(51.507765, 49.54621, 48.107232, 38.889494, 0, -37.11174, -53.104781,
         -0.127924, 18.398562, -122.778671, -77.035242, 0, -12.28863, 73.517283),
        ncol = 2,
        dimnames = list(
            c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument",
              "Null Island", "Tristan da Cunha", "Mawson Peak")
        )
    ))
#>                          [,1]        [,2]
#> Nelson's Column      51.50776   -0.127924
#> Ostravice            49.54621   18.398562
#> Tally Ho             48.10723 -122.778671
#> Washington Monument  38.88949  -77.035242
#> Null Island           0.00000    0.000000
#> Tristan da Cunha    -37.11174  -12.288630
#> Mawson Peak         -53.10478   73.517283

latlon(ll_mtx, decimal = TRUE)
#> $`Nelson's Column`
#> 	51.507765, -0.127924 decimal degrees
#> 
#> $Ostravice
#> 	49.546210, 18.398562 decimal degrees
#> 
#> $`Tally Ho`
#> 	48.107232, -122.778671 decimal degrees
#> 
#> $`Washington Monument`
#> 	38.889494, -77.035242 decimal degrees
#> 
#> $`Null Island`
#> 	0.000000, 0.000000 decimal degrees
#> 
#> $`Tristan da Cunha`
#> 	-37.111740, -12.288630 decimal degrees
#> 
#> $`Mawson Peak`
#> 	-53.104781, 73.517283 decimal degrees
#> 

## Two-column numeric `matrix` in degrees, minutes and seconds
## Deg Min Sec
ll_mtx[1:14] <- c(
    51.3027954, 49.3246368, 48.0626035, 38.5322178, 0, -37.0642264, -53.0617212,
    -0.0740526, 18.2354822, -122.464322, -77.0206871, 0, -12.1719068, 73.3102219
)

ll_mtx
#>                          [,1]         [,2]
#> Nelson's Column      51.30280   -0.0740526
#> Ostravice            49.32464   18.2354822
#> Tally Ho             48.06260 -122.4643220
#> Washington Monument  38.53222  -77.0206871
#> Null Island           0.00000    0.0000000
#> Tristan da Cunha    -37.06423  -12.1719068
#> Mawson Peak         -53.06172   73.3102219

latlon(ll_mtx)
#> $`Nelson's Column`
#> 	51°30'27.954"N, 0°07'40.526"W
#> 
#> $Ostravice
#> 	49°32'46.368"N, 18°23'54.822"E
#> 
#> $`Tally Ho`
#> 	48°06'26.035"N, 122°46'43.220"W
#> 
#> $`Washington Monument`
#> 	38°53'22.178"N, 77°02'06.871"W
#> 
#> $`Null Island`
#> 	0°00'00.000"N, 0°00'00.000"E
#> 
#> $`Tristan da Cunha`
#> 	37°06'42.264"S, 12°17'19.068"W
#> 
#> $`Mawson Peak`
#> 	53°06'17.212"S, 73°31'02.219"E
#> 

rm(ll_mtx)