Optimise the Value of Lambda for a Box-Cox Transformation

Optimise the value of \(\lambda\) for a Box-Cox transformation by minimising skewness and return the optimally transformed data together with the optimal \(\lambda\) and corresponding skewness minimum as attributes.

Usage

opt_bc(x, interval = c(-5, 5), tol = 1e-10)

Arguments

x

a numeric vector of data to be transformed using optimised values.

interval

a vector containing the end-points of the interval to be searched for the minimum.

tol

the desired accuracy.

Value

A numeric vector the same length as x with attributes

lambda and skew representing the optimal value of \(\lambda\) and corresponding skewness minimum respectively.

Details

Uses the functions boxcox3 to create a Box-Cox transformation function for the given dataset, skewness to calculate the skewness of the transformed data and optimise to search for the value of \(\lambda\) that minimises skewness.

See also

Examples

runif(20) |> opt_bc()
#>  [1] -0.15597288 -0.53365083 -1.23612480 -0.93339906 -0.66678514 -0.81683315
#>  [7] -0.60798614 -0.01444385 -0.81501059 -0.77462208 -1.02784796 -1.30040049
#> [13] -0.79745251 -0.60482497 -0.09510107 -0.49164626 -0.21057974 -1.31122392
#> [19] -0.61618332 -0.09649962
#> attr(,"skew")
#> [1] 4.8654e-09
#> attr(,"lambda")
#> [1] 0.6259281
rlnorm(20) |> opt_bc()
#>  [1] -1.68620053  0.88661906  0.30002174  0.99722077  1.42194458 -0.48861425
#>  [7] -0.18026082  0.94158837  2.32627183  0.70699828 -1.81276730 -0.68843502
#> [13]  0.70233924 -1.34046702  2.24483180 -0.54113439 -0.34673489  0.24303299
#> [19] -2.12227173 -0.07127134
#> attr(,"skew")
#> [1] 5.803662e-10
#> attr(,"lambda")
#> [1] -0.0493604
rpois(100, 3) |> opt_bc()         ## Provokes warning
#> Warning: NA/Inf replaced by maximum positive value
#>   [1] -1.4109158  1.6628114  1.6628114  0.8950824  2.3580035  0.0000000
#>   [7]  0.0000000 -1.4109158  3.0037998  2.3580035  2.3580035  0.0000000
#>  [13]  1.6628114  1.6628114  3.0037998  1.6628114  0.8950824  1.6628114
#>  [19]  1.6628114  0.0000000  1.6628114 -1.4109158  0.8950824  1.6628114
#>  [25] -1.4109158  3.6127784  0.8950824  0.8950824  0.8950824  1.6628114
#>  [31]  2.3580035  2.3580035  0.8950824  2.3580035  1.6628114  0.8950824
#>  [37]  1.6628114  1.6628114  3.0037998  4.1927518  1.6628114  0.8950824
#>  [43] -1.4109158  1.6628114  0.0000000  1.6628114  0.8950824  1.6628114
#>  [49]  0.8950824  1.6628114  5.2853017  0.8950824  2.3580035  0.0000000
#>  [55]  2.3580035  1.6628114  0.8950824  0.0000000  1.6628114  2.3580035
#>  [61]  2.3580035  1.6628114  0.0000000  0.8950824 -1.4109158  2.3580035
#>  [67]  0.8950824  0.0000000 -1.4109158  2.3580035  2.3580035  0.0000000
#>  [73]  4.7489991  4.7489991  0.8950824  0.0000000 -1.4109158  0.8950824
#>  [79]  3.0037998  2.3580035  1.6628114  3.0037998  0.8950824  3.0037998
#>  [85]  0.8950824  0.8950824  2.3580035  2.3580035  0.8950824  2.3580035
#>  [91]  3.6127784  0.0000000  0.8950824  2.3580035  0.8950824  2.3580035
#>  [97]  3.0037998  1.6628114  0.0000000  1.6628114
#> attr(,"skew")
#> [1] 2.608833e-08
#> attr(,"lambda")
#> [1] 0.7087595
rpois(100, 3) |> opt_bc(c(-3, 5)) ## Redefining interval avoids warning
#>   [1]  2.5706429  1.5021528  3.0296312  0.0000000  2.0673075  0.8415133
#>   [7]  0.8415133  0.8415133  2.5706429  0.8415133  2.5706429  0.0000000
#>  [13]  2.0673075  2.0673075  0.0000000  3.0296312  1.5021528  0.8415133
#>  [19]  4.2287953  1.5021528  1.5021528  2.0673075  0.0000000  2.5706429
#>  [25]  2.0673075  0.0000000  3.0296312  1.5021528  1.5021528  2.0673075
#>  [31]  2.0673075  2.0673075  1.5021528  2.0673075  1.5021528  1.5021528
#>  [37]  2.0673075  3.0296312  0.0000000  2.5706429  1.5021528  1.5021528
#>  [43]  0.0000000  2.0673075  0.0000000  0.8415133  2.0673075  0.8415133
#>  [49]  2.5706429  0.8415133  2.0673075  3.8528665  0.0000000  1.5021528
#>  [55]  2.5706429  3.8528665  2.5706429  1.5021528  4.2287953  0.0000000
#>  [61]  3.0296312  1.5021528  3.0296312  2.0673075  0.0000000  1.5021528
#>  [67]  0.8415133  2.0673075  0.8415133  0.8415133  3.4547471  1.5021528
#>  [73]  2.0673075  0.8415133  0.8415133  2.0673075  0.0000000  0.0000000
#>  [79]  0.0000000  1.5021528  0.8415133  0.0000000  0.0000000  0.8415133
#>  [85]  0.8415133  2.0673075 -1.8427782  2.0673075  2.0673075  2.0673075
#>  [91]  0.0000000  3.4547471  1.5021528  1.5021528  1.5021528  2.0673075
#>  [97]  2.0673075  0.8415133  0.8415133  2.0673075
#> attr(,"skew")
#> [1] 7.894981e-09
#> attr(,"lambda")
#> [1] 0.5426589